### 2024-09-26 #7293 @kyleniemeyer
Hello @dagghe, I see in the submission notes that your original submission of pyOMA was rejected from JOSS (https://github.com/openjournals/joss-reviews/issues/3264), but was published at SoftwareX.

Can you provide more detail on how this version differs from the original?
