### 2024-09-26 #7293 @dagghe
Hi @kyleniemeyer, the old module was more a collection of functions with a gui, while the new module leverages classes flexibility so to offer many more functionalities such as the ability to process multi-setup experiments, to have interactive plots to select the poles (or peaks) to extract, the possibility to define the geometry of the tested structure and plot/animate the mode shapes, storing and retrieving results and more. You can check by yourself if you want to, new repo: https://github.com/dagghe/pyOMA2
Old repo: https://github.com/dagghe/PyOMA
