### 2024-09-28 e936f8e
add: pre commit in github action
Former-commit-id: 987194b9fe3a58d1cd48c54d785bf1cca608c851 [formerly 4d8e4759718218346c60e104d8c9ba2ff2b10537] [formerly 3fdead9582758f24b55b14ac3bc9162a42c1d237 [formerly cc9ff4f842f12328b6e4c37c4e5ec5f492113a82]]
Former-commit-id: 67e5e918cb29ae5b0b34e57ca8dc8c054083e5fd [formerly 206dea41a5feaa974e659f7010b55ab890f63f7a]
Former-commit-id: 5728b1fd41d10f641fe1a1c7f606191276df8ae0
