### 2024-10-15 aa2d0b1
add: DOI in paper.bib
Former-commit-id: b20036a2338cdd7672b917d04921f60cf343da8c [formerly 55f5115fd28f7e47c539f720ad2087496fa8a797] [formerly fff1ddb2d0b0f2f3ac5ff011588cd1394c9b49f8 [formerly 8622798171188d7b6396c3352ce9ea083ba02582]]
Former-commit-id: ccaa820fc703e3b434528e4851ae01623cec24bf [formerly 796e42cd541a9958b075289e65cea2b08385cc5d]
Former-commit-id: 39a7a5daf30c347a0bd95d632da847ad74473ff4
