+++ paper/paper.bib
@@ -5,7 +5,9 @@
   volume={508},
   pages={116207},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jsv.2021.116207},
+  url={https://doi.org/10.1016/j.jsv.2021.116207},
 }
 
 @article{peeters1999reference,
@@ -16,7 +18,9 @@
   number={6},
   pages={855--878},
   year={1999},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {10.1006/mssp.1999.1249},
+  url = {https://doi.org/10.1006/mssp.1999.1249},
 }
 
 @phdthesis{dohler2011subspace,
@@ -34,7 +38,9 @@
   number={2},
   pages={346--366},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {1016/j.ymssp.2013.01.012},
+  url = {https://doi.org/10.1016/j.ymssp.2013.01.012},
 }
 
 @article{peeters2004polymax,
@@ -45,7 +51,9 @@
   number={3-4},
   pages={395--409},
   year={2004},
-  publisher={IOS Press}
+  publisher={IOS Press},
+  doi = {10.1155/2004/523692},
+  url = {https://doi.org/10.1155/2004/523692},
 }
 
 @article{brincker2001modal,
@@ -76,7 +84,9 @@
   number={5},
   pages={1227--1239},
   year={2010},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ymssp.2009.10.024},
+  url={https://doi.org/10.1016/j.ymssp.2009.10.024},
 }
 
 @book{brincker2015introduction,
@@ -93,7 +103,9 @@
   volume={142},
   pages={143},
   year={2014},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/978-1-4939-0767-0},
+  url={https://doi.org/10.1007/978-1-4939-0767-0},
 }
 
 @article{reynders2012system,
@@ -103,7 +115,9 @@
   volume={19},
   pages={51--124},
   year={2012},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s11831-012-9069-x},
+  url={https://doi.org/10.1007/s11831-012-9069-x},
 }
 
 @book{van2012subspace,
@@ -121,7 +135,9 @@
   number={2},
   pages={562--581},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.ymssp.2012.11.011},
+  url={https://doi.org/10.1016/j.ymssp.2012.11.011},
 }
 
 
@@ -134,6 +150,8 @@
   pages={121069},
   year={2021},
   publisher={Elsevier},
+  doi={10.1016/j.conbuildmat.2020.121069},
+  url={https://doi.org/10.1016/j.conbuildmat.2020.121069},
 }
 
 @article{aloisio2020dynamic,
@@ -143,7 +161,9 @@
   volume={213},
   pages={110593},
   year={2020},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.engstruct.2020.110593},
+  url={https://doi.org/10.1016/j.engstruct.2020.110593},
 }
 
 @article{simoncelli2023intensity,
@@ -153,7 +173,9 @@
   volume={206},
   pages={107937},
   year={2023},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.jcsr.2023.107937},
+  url={https://doi.org/10.1016/j.jcsr.2023.107937},
 }
 
 @article{saharan2023convolutional,
@@ -162,7 +184,9 @@
   journal={Journal of Vibration and Control},
[…]
