### 2024-11-01 #7293 @dagghe
Hi @faroit, I had a look at the JOSS reviewers list and think that the following user could be a fit based on their background: e-dub, Nitnelav, joaoadelazzari, CWillberg, arunmano121.
