### 2024-11-14 44844ba
Add test for testing plot_data for single column data
