### 2024-11-18 4199545
fix:uncertainties calculations
Former-commit-id: ba6e756a3323908f0bd8822c47138cb7dc0f2862 [formerly 3a1b8cd8e498a40f5094222dda10b2320e6955f9] [formerly f5eea7bbefa694e50ee83513a59e02c3dc809876 [formerly 819135d21185ddda70f16e3d908d87022e8df7b7]]
Former-commit-id: ac6d55b112ed7f3ca0e28341d84ca62c8ca8c7c4 [formerly 44a3df8433465eacc5d3b59305696f4db3ba002c]
Former-commit-id: 35b688498e2bd6cdbd4508266d0a5dc712c5c1cb
