+++ src/pyoma2/functions/ssi.py
@@ -28,7 +28,7 @@ def build_hank(
     method: typing.Literal["cov_mm", "cov_R", "dat"],
     calc_unc: bool = False,
     nb: int = 100,
-) -> typing.Tuple[np.ndarray, typing.Union[None, np.ndarray]]:
+) -> typing.Tuple[np.ndarray, typing.Optional[np.ndarray]]:
     """
     Construct a Hankel matrix using various methods with optional uncertainty calculations.
 
@@ -42,185 +42,123 @@ def build_hank(
         The number of block rows to use in the Hankel matrix.
     method : str
         The method to use for constructing the Hankel matrix.
-        Options are 'cov_mm', 'cov_R', 'dat'ì
+        Options are 'cov_mm', 'cov_R', 'dat'
     calc_unc : bool, optional
         Whether to calculate uncertainties (default is False). Only applicable for the 'cov_mm' method.
     nb : int, optional
-        The number of bootstrap samples to use for uncertainty calculations (default is 100).
+        The number of data blocks to use for uncertainty calculations (default is 100).
 
     Returns
     -------
     Hank : np.ndarray
         The constructed Hankel matrix.
     T : np.ndarray, optional
-        The uncertainty matrix. Only returned if `calc_unc` is True and `method` is 'cov_mm'.
+        The uncertainty matrix. Only returned if `calc_unc` is True.
 
     Raises
     ------
     AttributeError
-        If `calc_unc` is True but `method` is not 'cov_mm'.
-    AttributeError
-        If `method` is not one of 'cov_mm', 'cov_R', 'dat'ì
-
-    Notes
-    -----
-    - The 'YfYp' method constructs separate future and past data matrices without combining them into a Hankel matrix.
-    - When `calc_unc` is True, the uncertainty matrix `T` is calculated using the bootstrap method, which quantifies the uncertainty of the Hankel matrix elements.
+        If `method` is not one of 'cov_mm', 'cov_R', 'dat'
     """
     Ndat = Y.shape[1]  # number of data points
-    l = Y.shape[0]  # number of chaiiels  # noqa E741 (ambiguous variable name 'l')
-    r = Yref.shape[0]  # number of reference chaiiels
-    p = int(br)  # number of block row (to use same notation as Dohler)
-    q = int(p + 1)  # block column
-    N = Ndat - p - q  # lenght of the Hankel matrix
+    l = Y.shape[0]  # number of channels # noqa E741 (ambiguous variable name 'l')
+    r = Yref.shape[0]  # number of reference channels
+    p = br - 1  # number of block rows
+    q = br  # block columns
+    N = Ndat - p - q  # length of the Hankel matrix
 
-    if calc_unc and method != "cov_mm":
-        raise AttributeError(
-            "Uncertainty calculations are only available for 'cov_mm' method"
-        )
-
+    T = None
     logger.info("Assembling Hankel matrix method: %s...", method)
+
     if method == "cov_mm":
         # Future and past Output (Y^+ and Y^-)
-        Yf = np.vstack([(1 / N**0.5) * Y[:, q + 1 + i : N + q + i] for i in range(p + 1)])
-        Yp = np.vstack(
-            [(1 / N**0.5) * Yref[:, q + i : N + q - 1 + i] for i in range(0, -q, -1)]
-        )
+        Yf = np.vstack([Y[:, q + i : N + q + i] for i in range(p + 1)])
[…]
