### 2024-11-18 3733e49
Add assertions to test
Former-commit-id: a416106980ce887eb98e78def1213c4dd5c64283 [formerly e811263f286c790a6f6f5f83f18912da3cb803c2] [formerly 6e51a5244f15576c839bd21f0196d88ccbd5cb88 [formerly 1e6e6b519ee765fffd4f456339abe7b5024b7149]]
Former-commit-id: 3fa7a206d63dd5a8b8fa8362cccbc5ba22f5fe91 [formerly 65a40fd053ba452f7f5c25ee3710bf2732f4e47e]
Former-commit-id: ba4053aaf5b9b319fa028d34f81f606d685da564
