+++ tests/unit/test_setups.py
@@ -278,6 +278,12 @@ def test_init_multisetup_preger(multi_setup_data_fixture) -> None:
 
 
 def test_plot_data_single_column(ss):
-    ss.data = ss.data[:, 0]
-    ss.plot_data()
-    assert False
+    # remove all but one column from data
+    ss.data = ss.data[:, 0][:, None]
+
+    _, axs = ss.plot_data()
+
+    # check for right dimensions
+    assert np.asarray(axs).ndim == 2
+    assert len(axs[0]) == 1
+    axs[0][0].plot.assert_called()
