### 2024-11-18 6937c6f
Fix axes dimension with 'squeeze=False'
Former-commit-id: 549a2c8ca6a03147007bd508648395a185159dd3 [formerly 4b708c59b50d3c3064c2573d5334681ab5b29a93] [formerly 6b6d7827e27be0b3cbe8efbb23e606d19279908b [formerly 630a7c763bd52d2094c0e5d2403283aff2bb11c6]]
Former-commit-id: 4800609571925df9383f315153db3939194dc3c8 [formerly 9941f63269ef4ed2eb8a39086581784d4157fa68]
Former-commit-id: 13251be7eb8e17f0de3d9375d433701ca615a4ba
