### 2024-11-18 e6d3b31
Adapt fixture to squeeze keyword
