### 2024-11-18 fb6fccc
Do not only return last axis, but all of them
Former-commit-id: 22e35945dd73b47463ff83d5031b62ab0b4f34be [formerly 473f4bd3a3135e2f755d50e83fae15a5bde7aa7d] [formerly a3a5fc69dbb2b433cd34d298b71329d6c7e04156 [formerly 484f5dc108147479ca653cb674535416735c27a5]]
Former-commit-id: 309fae4bb845db1b27db932004376b3c2b60943c [formerly d8b4e51771b4367e56639480ee04f5bb56f31dd6]
Former-commit-id: 7a73c2ce41b9836f1a7b81c439a16008d26ab9a3
