### 2024-11-18 aca961e
change: cov_mm to cov across all files, some default values in run_params; add: HC_mpc and HC_mpd functions; remove:HC_phi_comp function
Former-commit-id: fbc78600f08f6b40a26e648116ec456b5d7f3790 [formerly 60628be4e575530d65e48a1afa0bfc2ddc6b91ac] [formerly fe94b5dcd05df5ac3ccdc838687525a3f4008a3e [formerly 2b5f9d628278e72adccd3ac4824c29948a5c1eb8]]
Former-commit-id: 31eadca6fa950ca9c766c114d3d70e0539e7b706 [formerly f53d3d8a8e5d5ed1c9991973f480ece92da7c670]
Former-commit-id: 87eea7ffd60829a716a6524f969903ea37b1d19a
