+++ src/pyoma2/algorithms/ssi.py
@@ -109,18 +109,22 @@ class SSIdat(BaseAlgorithm[SSIRunParams, SSIResult, typing.Iterable[float]]):
         hc_CoV_max = hc["CoV_max"]
 
         # HC - MPC and MPD
-        mask3, mask4 = gen.HC_phi_comp(Phis, hc_mpc_lim, hc_mpd_lim)
-        lista = [Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std]
-        Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std = gen.applymask(
-            lista, mask3, Phis.shape[2]
-        )
-        Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std = gen.applymask(
-            lista, mask4, Phis.shape[2]
-        )
+        if hc_mpc_lim is not None:
+            mask3 = gen.HC_MPC(Phis, hc_mpc_lim)
+            lista = [Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std]
+            Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std = gen.applymask(
+                lista, mask3, Phis.shape[2]
+            )
+        if hc_mpd_lim is not None:
+            mask4 = gen.HC_MPD(Phis, hc_mpd_lim)
+            lista = [Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std]
+            Fns, Xis, Phis, Lambds, Fn_std, Xi_std, Phi_std = gen.applymask(
+                lista, mask4, Phis.shape[2]
+            )
 
         # HC - maximum covariance
         if Fn_std is not None and hc_CoV_max is not None:
-            Fn_std, mask5 = gen.HC_cov(Fns, Fn_std, hc_CoV_max)
+            Fn_std, mask5 = gen.HC_CoV(Fns, Fn_std, hc_CoV_max)
             lista = [Fns, Xis, Phis, Lambds, Xi_std, Phi_std]
             Fns, Xis, Phis, Lambds, Xi_std, Phi_std = gen.applymask(
                 lista, mask5, Phis.shape[2]
@@ -158,7 +162,7 @@ class SSIdat(BaseAlgorithm[SSIRunParams, SSIResult, typing.Iterable[float]]):
     def mpe(
         self,
         sel_freq: typing.List[float],
-        order: typing.Union[int, str] = "find_min",
+        order_in: typing.Union[int, str] = "find_min",
         rtol: float = 5e-2,
     ) -> typing.Any:
         """
@@ -179,11 +183,11 @@ class SSIdat(BaseAlgorithm[SSIRunParams, SSIResult, typing.Iterable[float]]):
         typing.Any
             The extracted modal parameters. The format and content depend on the algorithm's implementation.
         """
-        super().mpe(sel_freq=sel_freq, order=order, rtol=rtol)
+        super().mpe(sel_freq=sel_freq, order_in=order_in, rtol=rtol)
 
         # Save run parameters
         self.run_params.sel_freq = sel_freq
-        self.run_params.order_in = order
+        self.run_params.order_in = order_in
         self.run_params.rtol = rtol
 
         # Get poles
@@ -201,7 +205,7 @@ class SSIdat(BaseAlgorithm[SSIRunParams, SSIResult, typing.Iterable[float]]):
             Fn_pol,
             Xi_pol,
             Phi_pol,
-            order,
+            order_in,
             Lab=Lab,
             rtol=rtol,
             Fn_std=Fn_pol_std,
@@ -407,7 +411,7 @@ class SSIcov(SSIdat):
     Attributes
[…]
