### 2024-11-22 794cdd2
fix: input arguments in ssi algorithms; MPC and MPD functions to handle nan; plot_ch_info now return list of figs and axes; docstrings for ssi; input arguments and range of loops for ssi
Former-commit-id: 3e662d8fd299b04b7e4825bdf4af59e749f20849 [formerly 478100873404e207a17fd03f3553ef700fddffed] [formerly 55b3904ec053c5834566ae1ce3877a576e599148 [formerly 0aef4b3c0fe4f516ee5f7b7e77dd5e9f4568c617]]
Former-commit-id: 8973a792addfd0963f4c076adb64f8b1abae7f9e [formerly b7232dc66f2cd245602f89ba3f87255f03281299]
Former-commit-id: f2fe7fd4b0d649746e41a7f00c3f22d898d01f16
