+++ src/pyoma2/functions/ssi.py
@@ -27,10 +27,13 @@ def build_hank(
     br: int,
     method: typing.Literal["cov", "cov_R", "dat"],
     calc_unc: bool = False,
-    nb: int = 100,
+    nb: int = 50,
 ) -> typing.Tuple[np.ndarray, typing.Optional[np.ndarray]]:
     """
-    Construct a Hankel matrix using various methods with optional uncertainty calculations.
+    Construct a Hankel matrix with optional uncertainty calculations.
+
+    This function constructs a Hankel matrix using various methods, such as covariance-based
+    or data-driven approaches, and optionally computes uncertainty matrices.
 
     Parameters
     ----------
@@ -39,27 +42,35 @@ def build_hank(
     Yref : np.ndarray
         The reference data matrix with shape (number of reference channels, number of data points).
     br : int
-        The number of block rows to use in the Hankel matrix.
-    method : str
-        The method to use for constructing the Hankel matrix.
-        Options are 'cov', 'cov_R', 'dat'
+        The number of block rows in the Hankel matrix.
+    method : {'cov', 'cov_R', 'dat'}
+        The method for constructing the Hankel matrix:
+        - 'cov': Covariance-based.
+        - 'cov_R': Covariance-based with auto-correlations.
+        - 'dat': Data-driven.
     calc_unc : bool, optional
-        Whether to calculate uncertainties (default is False). Only applicable for the 'cov' method.
+        Whether to calculate uncertainty matrices. Default is False.
     nb : int, optional
-        The number of data blocks to use for uncertainty calculations (default is 100).
+        The number of data blocks to use for uncertainty calculations. Default is 50.
 
     Returns
     -------
     Hank : np.ndarray
         The constructed Hankel matrix.
-    T : np.ndarray, optional
-        The uncertainty matrix. Only returned if `calc_unc` is True.
+    T : np.ndarray or None
+        The uncertainty matrix, returned only if `calc_unc` is True. None otherwise.
 
     Raises
     ------
     AttributeError
-        If `method` is not one of 'cov', 'cov_R', 'dat'
+        If `method` is not one of {'cov', 'cov_R', 'dat'}, or if data is insufficient for
+        uncertainty calculation.
+
+    Notes
+    -----
+    Uncertainty calculations follow the approach detailed in
     """
+
     Ndat = Y.shape[1]  # number of data points
     l = Y.shape[0]  # number of channels # noqa E741 (ambiguous variable name 'l')
     r = Yref.shape[0]  # number of reference channels
@@ -91,7 +102,13 @@ def build_hank(
                 Hcov_k = np.dot(Yf_k, Yp_k.T) / Nb
 
                 Hcov_vec_k = Hcov_k.reshape(-1, order="F")
-                T[:, k] = (Hcov_vec_k - Hvec0) / np.sqrt(nb * (nb - 1))
+                if Hcov_vec_k.shape[0] < T.shape[0]:
+                    raise AttributeError(
+                        "Not enough data points per data block."
+                        "Try reducing the number of data blocks, nb and/or the number of block-rows, br"
+                    )
+                else:
+                    T[:, k] = (Hcov_vec_k - Hvec0) / np.sqrt(nb * (nb - 1))
 
     elif method == "cov_R":
         # Correlations
@@ -122,7 +139,13 @@ def build_hank(
                 )
 
                 Hcov_vec_j = Hcov_j.reshape(-1, 1, order="f")
-                T[:, j - 1] = (Hcov_vec_j - Hvec0).flatten() / np.sqrt(nb * (nb - 1))
+                if Hcov_vec_j.shape[0] < T.shape[0]:
+                    raise AttributeError(
+                        "Not enough data points per data block."
[…]
