### 2024-11-23 27d80dc
Update issue templates
