+++ .github/ISSUE_TEMPLATE/custom.md
@@ -1,0 +1,32 @@
+---
+name: Custom issue template
+about: Describe this issue template's purpose here.
+title: ''
+labels: ''
+assignees: ''
+
+---
+
+**Describe the bug**
+A clear and concise description of what the bug is.
+
+**To Reproduce**
+Steps to reproduce the behavior:
+1. Go to '...'
+2. Click on '....'
+3. Scroll down to '....'
+4. See error
+
+**Expected behavior**
+A clear and concise description of what you expected to happen.
+
+**Screenshots**
+If applicable, add screenshots to help explain your problem.
+
+**Desktop (please complete the following information):**
+ - pyoma2 version: [e.g. 0.1.0 or latest]
+ - OS: [e.g. iOS]
+ - Version [e.g. 22]
+
+**Additional context or possible solutions**
+Add any other context about the problem here.
