### 2024-11-26 2952bed
change: enhanced dd.plot_EFDDfit error log
