### 2024-11-26 12369ff
change: enhanced dd.plot_EFDDfit error log
Former-commit-id: 01f69623fafe76be6a7fccc3aa23f5c79739c80e [formerly c4051bb8baa723cc192e263b33e993e5f2f9c569] [formerly b4768af63aaa370e2e506b961951900c7d1d5a69 [formerly 2952bed282aa0e1532d04b821c4508d722a2381c]]
Former-commit-id: 58153996f55b0d2a10d762f7de8592677ef38989 [formerly 35ee8fc508a4dd2fb71e6a7f41f727eba7d33631]
Former-commit-id: 53cadb3efe01944a0d2dfc1aaef8b1257771f49c
