### 2024-12-03 7ae6639
Fix plot cluster for pLSCF
Maybe old code variable name Sm that has to be changed to Xi in order for the clusters to be plotted with pLSCF
