### 2024-12-07 #7293 @dagghe
Hi @faroit, I noticed that e-dub gave your request a thumbs-up but didn’t reply. Others I know with expertise in both Python and structural dynamics are Junko Slavic (from the University of Ljubljana) and Gunnstein Thomas Frøseth (from the Norwegian University of Science and Technology). However, I’m currently working on a STAR paper with both of them, so I’m not sure if that might be considered a conflict of interest?!
I could suggest other peers in the field, but I’m not sure how strong they are with Python or whether they even have GitHub accounts.
