+++ src/pyoma2/algorithms/data/result.py
@@ -5,13 +5,11 @@ related to the pyOMA2 module.
 
 from __future__ import annotations
 
-import typing
+from typing import List, Optional, Union
 
 import numpy as np
 import numpy.typing as npt
 from pydantic import BaseModel, ConfigDict
-
-"""TODO fix type"""
 
 
 class BaseResult(BaseModel):
@@ -20,16 +18,16 @@ class BaseResult(BaseModel):
 
     Attributes
     ----------
-    Fn : numpy.typing.NDArray
+    Fn : numpy.NDArray
         Array of natural frequencies obtained from modal analysis.
-    Phi : numpy.typing.NDArray
+    Phi : numpy.NDArray
         Array of mode shape vectors obtained from modal analysis.
     """
 
     model_config = ConfigDict(from_attributes=True, arbitrary_types_allowed=True)
     # dopo mpe o mpe_from_plot
-    Fn: typing.Optional[npt.NDArray[np.float64]] = None  # array of natural frequencies
-    Phi: typing.Optional[npt.NDArray[np.float64]] = None  # array of Mode shape vectors
+    Fn: Optional[npt.NDArray[np.float64]] = None  # array of natural frequencies
+    Phi: Optional[npt.NDArray[np.float64]] = None  # array of Mode shape vectors
 
 
 class FDDResult(BaseResult):
@@ -38,20 +36,20 @@ class FDDResult(BaseResult):
 
     Attributes
     ----------
-    freq : numpy.typing.NDArray
+    freq : numpy.NDArray
         Array of frequencies.
-    Sy : numpy.typing.NDArray
+    Sy : numpy.NDArray
         PSD obtained from the FDD analysis.
-    S_val : numpy.typing.NDArray
+    S_val : numpy.NDArray
         Singular values of the PSD.
-    S_vec : numpy.typing.NDArray
+    S_vec : numpy.NDArray
         Singular vectors of the PSD.
     """
 
-    freq: typing.Optional[npt.NDArray[np.float64]] = None
-    Sy: typing.Optional[npt.NDArray[np.float64]] = None
-    S_val: typing.Optional[npt.NDArray[np.float64]] = None
-    S_vec: typing.Optional[npt.NDArray[np.float64]] = None
+    freq: Optional[npt.NDArray[np.float64]] = None
+    Sy: Optional[npt.NDArray[np.float64]] = None
+    S_val: Optional[npt.NDArray[np.float64]] = None
+    S_vec: Optional[npt.NDArray[np.float64]] = None
 
 
 class EFDDResult(FDDResult):
@@ -61,23 +59,23 @@ class EFDDResult(FDDResult):
 
     Attributes
     ----------
-    freq : numpy.typing.NDArray
+    freq : numpy.NDArray
         Array of frequencies.
-    Sy : numpy.typing.NDArray
+    Sy : numpy.NDArray
         PSD obtained from the analysis.
-    S_val : numpy.typing.NDArray
+    S_val : numpy.NDArray
         Singular values of the PSD.
-    S_vec : numpy.typing.NDArray
+    S_vec : numpy.NDArray
         Singular vectors of the PSD.
-    Xi : numpy.typing.NDArray
+    Xi : numpy.NDArray
         Array of damping ratios obtained from modal analysis.
     forPlot : list
         A list to store data for plotting purposes.
     """
 
     # dopo mpe, MPE_forPlot
-    Xi: typing.Optional[npt.NDArray[np.float64]] = None  # array of damping ratios
-    forPlot: typing.Optional[typing.List] = None
+    Xi: Optional[npt.NDArray[np.float64]] = None  # array of damping ratios
+    forPlot: Optional[List] = None
 
 
 class SSIResult(BaseResult):
@@ -86,65 +84,61 @@ class SSIResult(BaseResult):
 
[…]
