+++ src/pyoma2/functions/clus.py
@@ -1,0 +1,1707 @@
+#!/usr/bin/env python3
+# -*- coding: utf-8 -*-
+"""
+Created on Sun Nov 24 07:07:42 2024
+
+@author: dagghe
+"""
+
+import numpy as np
+import pandas as pd
+from scipy import signal, stats
+from sklearn.cluster import (
+    HDBSCAN,
+    OPTICS,
+    AffinityPropagation,
+    AgglomerativeClustering,
+    KMeans,
+    SpectralClustering,
+)
+from sklearn.metrics.pairwise import pairwise_distances
+from sklearn.mixture import GaussianMixture
+from sklearn.preprocessing import power_transform
+
+from pyoma2.functions import gen
+
+# =============================================================================
+# CLUSTERING
+# =============================================================================
+
+
+def kmeans(feat_arr):
+    """
+    Perform k-means clustering on the given feature array.
+
+    Parameters
+    ----------
+    feat_arr : ndarray of shape (n_samples, n_features)
+        Input feature array for clustering.
+
+    Returns
+    -------
+    labels_all : ndarray of shape (n_samples,)
+        Cluster labels for each sample. Labels are adjusted such that the first cluster
+        corresponds to the smaller centroid (stable modes).
+    """
+    kmean = KMeans(n_clusters=2)
+    kmean.fit(feat_arr)
+    labels_all = kmean.labels_
+    # check the centroids to establish stable and spurious modes
+    centroids = kmean.cluster_centers_
+    # if the first centroid is larger than the second invert the labels
+    if centroids[0, 0] > centroids[1, 0]:
+        labels_all = 1 - labels_all
+    return labels_all
+
+
+# -----------------------------------------------------------------------------
+
+
+def GMM(feat_arr, dist=False):
+    """
+    Perform Gaussian Mixture Model (GMM) clustering on the given feature array.
+
+    Parameters
+    ----------
+    feat_arr : ndarray of shape (n_samples, n_features)
+        Input feature array for clustering.
+
+    Returns
+    -------
+    labels_all : ndarray of shape (n_samples,)
+        Cluster labels for each sample. Labels are adjusted such that the first cluster
+        corresponds to the smaller mean (stable modes).
+    """
+    GMM = GaussianMixture(n_components=2)
+    GMM.fit(feat_arr)
+    labels_all = GMM.predict(feat_arr)
+    # check the means to establish stable and spurious modes
+    centroids = GMM.means_
+    # if the first centroid is larger than the second invert the labels
+    if centroids[0, 0] > centroids[1, 0]:
+        labels_all = 1 - labels_all
+    if dist:
+        # calculate limit distance
+        pi0, pi1 = GMM.weights_
+        mu = GMM.means_
+        # sig = GMM.covariances_
+        dlim = pi0 * mu[0, 0] + pi1 * mu[1, 1]
+        return labels_all, dlim
+    else:
+        return labels_all
+
+
+# -----------------------------------------------------------------------------
+
+
+def hierarc(dtot, dc, linkage, n_clusters, ordmax, step, Fns, Phis):
+    """
+    Perform hierarchical clustering with specified parameters.
+
+    Parameters
+    ----------
+    dtot : ndarray of shape (n_samples, n_samples)
+        Pairwise distance matrix for clustering.
+    dc : float or str, optional
+        Distance threshold for clustering. Special string options include:
+        - "mu+2sig": Mean plus two standard deviations of distances.
+        - "95weib": 95th percentile of a Weibull distribution fit to the distances.
+        - "auto": Automatic threshold estimation based on KDE.
+    n_clusters : int or str, optional
[…]
