+++ src/pyoma2/functions/plot.py
@@ -12,7 +12,10 @@ import matplotlib
 import matplotlib.pyplot as plt
 import matplotlib.tri as mtri
 import numpy as np
-from scipy import signal
+from matplotlib.colors import to_rgba
+from matplotlib.lines import Line2D
+from matplotlib.ticker import MultipleLocator
+from scipy import signal, stats
 from scipy.interpolate import interp1d
 
 from .gen import MAC
@@ -22,6 +25,598 @@ logger = logging.getLogger(__name__)
 # =============================================================================
 # PLOT ALGORITMI
 # =============================================================================
+
+
+def plot_dtot_hist(dtot, bins="auto", sugg_co=True):
+    """
+    Plot a histogram of the total distance matrix with optional suggested cut-off distances.
+
+    This function plots a histogram of the values in the input distance matrix or vector `dtot`.
+    It overlays a kernel density estimate (KDE) and optionally indicates suggested cut-off distances
+    for clustering using single-linkage and average-linkage methods.
+
+    Parameters
+    ----------
+    dtot : ndarray
+        The input distance data. If a 2D array is provided, the upper triangular elements
+        (excluding the diagonal) are extracted. If a 1D array is provided, it is used as is.
+    bins : int or str, optional
+        The number of bins for the histogram. Defaults to "auto".
+    sugg_co : bool, optional
+        Whether to compute and display suggested cut-off distances for clustering.
+        Defaults to True.
+
+    Returns
+    -------
+    fig : matplotlib.figure.Figure
+        The figure object containing the plot.
+    ax : matplotlib.axes.Axes
+        The axes object for the histogram plot.
+    """
+    if dtot.ndim == 2:
+        # Extract upper triangular indices and values
+        upper_tri_indices = np.triu_indices_from(dtot, k=0)
+        x = dtot[upper_tri_indices]
+    elif dtot.ndim == 1:
+        x = dtot
+
+    xs = np.linspace(dtot.min(), dtot.max(), 500)
+
+    kde = stats.gaussian_kde(x)
+    # Evaluate the KDE to get the PDF
+    pdf = kde(xs)
+
+    # Create a figure and axis
+    fig, ax = plt.subplots(figsize=(10, 6))
+
+    # Plot the histogram on the axis
+    ax.hist(x, bins=bins, color="skyblue", edgecolor="black")
+    ax1 = ax.twinx()
+    ax1.plot(xs, kde(xs), "k-", label="Kernel density estimate")
+
+    if sugg_co:
+        minima_in = signal.argrelmin(pdf)[0]
+        minima = pdf[minima_in]
+        min_abs = minima.argmin()
+        min_abs_ind = minima_in[min_abs]
+        maxima_in = signal.argrelmax(pdf)
+        dc2_ind = maxima_in[0][0]
+        dc2 = xs[dc2_ind]
+        dc1 = xs[min_abs_ind]
+
+        ax1.axvline(
+            dc2,
+            color="red",
+            linestyle="dashed",
+            linewidth=2,
+            label=f"Suggested cut-off distance single-linkage: {dc2:.4f}",
+        )
+        ax1.axvline(
+            dc1,
+            color="red",
+            linestyle="dotted",
+            linewidth=2,
+            label=f"Suggested cut-off distance average-linkage: {dc1:.4f}",
+        )
+
+    # Customize plot
+    ax.set_xlabel("dtot")
+    ax.set_ylabel("Frequency")
+    ax.set_title("Histogram of distances")
+    ax.xaxis.set_major_locator(MultipleLocator(0.1))  # Major ticks every 0.1
+    ax.xaxis.set_minor_locator(
+        MultipleLocator(0.02)
+    )  # Minor ticks every 1/5 of major (0.02)
+
+    # Add grid only for the x-axis
[…]
