+++ src/pyoma2/algorithms/data/run_params.py
@@ -6,10 +6,25 @@ algorithms included in the pyOMA2 module.
 from __future__ import annotations
 
 import typing
+from typing import List, Literal, Optional, Tuple, Union
 
 import numpy as np
 import numpy.typing as npt
-from pydantic import BaseModel, ConfigDict
+from pydantic import BaseModel, ConfigDict, model_validator
+from typing_extensions import TypedDict
+
+
+class HCDictType(TypedDict):
+    xi_max: Optional[float] = None
+    mpc_lim: Optional[float] = None
+    mpd_lim: Optional[float] = None
+    CoV_max: Optional[float] = None
+
+
+class SCDictType(TypedDict):
+    err_fn: float
+    err_xi: float
+    err_phi: float
 
 
 class BaseRunParams(BaseModel):
@@ -17,7 +32,9 @@ class BaseRunParams(BaseModel):
     Base class for storing run parameters for modal analysis algorithms.
     """
 
-    model_config = ConfigDict(from_attributes=True, arbitrary_types_allowed=True)
+    model_config = ConfigDict(
+        from_attributes=True, arbitrary_types_allowed=True, extra="forbid"
+    )
 
 
 class FDDRunParams(BaseRunParams):
@@ -44,10 +61,10 @@ class FDDRunParams(BaseRunParams):
 
     # METODO 1: run
     nxseg: int = 1024
-    method_SD: typing.Literal["per", "cor"] = "per"
+    method_SD: Literal["per", "cor"] = "per"
     pov: float = 0.5
     # METODO 2: mpe e mpe_from_plot
-    sel_freq: typing.Optional[npt.NDArray[np.float64]] = None
+    sel_freq: Optional[npt.NDArray[np.float64]] = None
     DF: float = 0.1
 
 
@@ -85,10 +102,10 @@ class EFDDRunParams(BaseRunParams):
 
     # METODO 1: run
     nxseg: int = 1024
-    method_SD: typing.Literal["per", "cor"] = "per"
+    method_SD: Literal["per", "cor"] = "per"
     pov: float = 0.5
     # METODO 2: mpe e mpe_from_plot
-    sel_freq: typing.Optional[npt.NDArray[np.float64]] = None
+    sel_freq: Optional[npt.NDArray[np.float64]] = None
     DF1: float = 0.1
     DF2: float = 1.0
     cm: int = 1
@@ -150,17 +167,17 @@ class SSIRunParams(BaseRunParams):
     # METODO 1: run
     br: int = 20
     method: str = None
-    ref_ind: typing.Optional[typing.List[int]] = None
+    ref_ind: Optional[List[int]] = None
     ordmin: int = 0
-    ordmax: typing.Optional[int] = None
+    ordmax: Optional[int] = None
     step: int = 1
-    sc: dict = dict(err_fn=0.01, err_xi=0.05, err_phi=0.03)  # soft criteria
-    hc: dict = dict(xi_max=0.1, mpc_lim=0.7, mpd_lim=0.3, CoV_max=0.02)  # hard criteria
+    sc: SCDictType = dict(err_fn=0.05, err_xi=0.05, err_phi=0.05)
+    hc: HCDictType = dict(xi_max=0.1, mpc_lim=0.5, mpd_lim=0.5, CoV_max=0.05)
     calc_unc: bool = False  # uncertainty calculations
     nb: int = 50  # number of dataset blocks
     # METODO 2: mpe e mpe_from_plot
-    sel_freq: typing.Optional[typing.List[float]] = None
-    order_in: typing.Union[int, list, str] = "find_min"
+    sel_freq: Optional[List[float]] = None
[…]
