### 2024-12-12 #7293 @faroit
> Others I know with expertise in both Python and structural dynamics are Janko Slavic (from the University of Ljubljana) and Gunnstein Thomas Frøseth (from the Norwegian University of Science and Technology). However, I’m currently working on a STAR paper with both of them, so I’m not sure if that might be considered a conflict of interest?!

@dagghe thanks for flagging this. This would indeed be a conflict of interest wrt to the JOSS policy https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy

Happy to look further now finding more suitable reviewers. Let me know if you have more suggestions on your side as well
