### 2024-12-12 #7293 @dagghe
@faroit I did a quick search on GitHub and these are my suggestions:
- The maintainers of the sdynpy repo (https://github.com/sandialabs/sdynpy)
- twmacro (https://github.com/twmacro/pyyeti)
- oiseth or someone from (https://github.com/oiseth/TKT4108StructuralDynamics2)
- knutankv (https://github.com/knutankv/koma)
