### 2025-01-05 f399138
fix: check A size in unc calc and save access to plscf conj
