### 2025-01-11 #7656 @editorialbot
[REVIEW]: pyOMA2: a Python module to conduct operational modal analysis
**Submitting author:** @dagghe (<a href="http://orcid.org/0000-0002-3830-2835">Dag</a>)
**Repository:** https://github.com/dagghe/pyOMA2
**Branch with paper.md** (empty if default branch): Joss
**Version:** 1.0.0
