### 2025-01-18 #7293 @faroit
@dprohe I removed you from the list of reviewers because of the potential conflict of interest because of the "STAR" paper you are co-working with the @dagghe. I sincerely apologize for the confusions and hope to have you on board some other time. I saw that you started a review list already in the review issue already, no need to do any further actions now. The JOSS system should work fine with the updated reviewers list now. Thanks again for your help
