### 2025-01-22 c7401e5
removed unused lines of code (comments legacy functions etc), added plot_mode_complexity function
Former-commit-id: 55bad5a7228a14e95edfd5d41c76adf85db368b1 [formerly 0953ed97471110d03d136037c3fd36cdf149d12c] [formerly f93845c9fdcc0be42957ef22c4e03d490e1c4669 [formerly 2612c24e2baebbd1a02574f6e2194cfc69dd4d13]]
Former-commit-id: 4d642c54808e7bd255137f85e93d4bf50b0aa485 [formerly e9508265b63dd388c4d64d2a2f2f767f9c601e12]
Former-commit-id: 786589e77e3049bde95f7eddf40cdcfb8f39017e
