+++ README.md
@@ -11,7 +11,11 @@ _______________________
 This is the new and updated version of pyOMA module, a Python module designed for conducting operational modal analysis.
 With this update, we've transformed pyOMA from a basic collection of functions into a more sophisticated module that fully leverages the capabilities of Python classes.
 
-The module now supports analysis of both single and multi-setup data measurements, which includes handling multiple acquisitions with a mix of reference and roving sensors. We've also introduced interactive plots, allowing users to select desired modes for extraction directly from the plots generated by the algorithms. Additionally, a new feature enables users to define the geometry of the structures being tested, facilitating the visualization of mode shapes after modal results are obtained. Since version 1.0.x, the uncertainty bounds of the modal properties can also be estimated for the SSI family of algorithms. Finally, the underlying functions have been rigorously revised, resulting in significant enhancements and optimizations.
+The module now supports analysis of both single and multi-setup data measurements, which includes handling multiple acquisitions with a mix of reference and roving sensors. We've also introduced interactive plots, allowing users to select desired modes for extraction directly from the plots generated by the algorithms. Additionally, a new feature enables users to define the geometry of the structures being tested, facilitating the visualization of mode shapes after modal results are obtained. Since version 1.0.1, the uncertainty bounds of the modal properties can also be estimated for the SSI family of algorithms. Finally, the underlying functions have been rigorously revised, resulting in significant enhancements and optimizations.
+
+*22/01/2025:
+We have introduced introduced a specialised class for clustering purposes, *
+
 
 ## Documentation
 
