### 2025-01-22 2612c24
removed unused lines of code (comments legacy functions etc), added plot_mode_complexity function
