+++ src/pyoma2/functions/plot.py
@@ -619,6 +619,9 @@ def stab_clus_plot(
     return fig, ax
 
 
+# -----------------------------------------------------------------------------
+
+
 def CMIF_plot(
     S_val: np.ndarray,
     freq: np.ndarray,
@@ -670,8 +673,9 @@ def CMIF_plot(
         try:
             assert int(nSv) < S_val.shape[1]
         except Exception as e:
-            # DA SISTEMARE!!!
-            raise ValueError("ERROR") from e
+            raise ValueError(
+                f"ERROR: nSV must be less or equal to {S_val.shape[1]}. nSV={int(nSv)}"
+            ) from e
 
     for k in range(nSv):
         if k == 0:
@@ -932,144 +936,6 @@ def stab_plot(
     return fig, ax
 
 
-# LEGACY
-def Stab_plot(
-    Fn: np.ndarray,
-    Lab: np.ndarray,
-    step: int,
-    ordmax: int,
-    ordmin: int = 0,
-    freqlim: typing.Optional[typing.Tuple] = None,
-    hide_poles: bool = True,
-    fig: typing.Optional[plt.Figure] = None,
-    ax: typing.Optional[plt.Axes] = None,
-) -> typing.Tuple[plt.Figure, plt.Axes]:
-    """
-    Plot the stabilization chart for modal analysis.
-
-    This function creates a stabilization chart, which is a graphical representation used in
-    system identification to assess the stability of identified modes across different model
-    orders.
-
-    Parameters
-    ----------
-    Fn : ndarray
-        An array containing the frequencies for each model order and identification step.
-    Lab : ndarray
-        An array of labels indicating the stability status of each pole. Different numbers represent
-        different stability statuses such as stable pole, stable frequency, stable mode shape, etc.
-    step : int
-        The step size between model orders in the identification process.
-    ordmax : int
-        The maximum model order to be displayed on the plot.
-    ordmin : int, optional
-        The minimum model order to be displayed on the plot, by default 0.
-    freqlim : tuple of float, optional
-        A tuple defining the frequency limits for the plot. If None, includes all frequencies.
-        Default is None.
-    hide_poles : bool, optional
-        If True, only stable poles are plotted; if False, all types of poles are plotted.
-        Default is True.
-    fig : matplotlib.figure.Figure, optional
-        An existing matplotlib figure object to plot on. If None, a new figure is created.
-        Default is None.
-    ax : matplotlib.axes.Axes, optional
-        An existing axes object to plot on. If None, new axes are created on the provided
-        or new figure. Default is None.
-
-    Returns
-    -------
-    tuple
-        - fig : matplotlib.figure.Figure
-            The matplotlib figure object containing the plot.
-        - ax : matplotlib.axes.Axes
-            The axes object with the stabilization chart.
-
-    Notes
-    -----
-    The stabilization chart helps in identifying the number of physical modes and their
-    stability by observing how poles behave across different model orders. Stable poles are
-    typically considered as indicators of physical modes.
-    """
-    if fig is None and ax is None:
-        fig, ax = plt.subplots(figsize=(8, 6), tight_layout=True)
-
-    # Stable pole
-    a = np.where(Lab == 7, Fn, np.nan)
-
[…]
