+++ src/pyoma2/functions/ssi.py
@@ -25,12 +25,15 @@ def build_hank(
     Y: np.ndarray,
     Yref: np.ndarray,
     br: int,
-    method: typing.Literal["cov_mm", "cov_R", "dat"],
+    method: typing.Literal["cov", "cov_R", "dat"],
     calc_unc: bool = False,
-    nb: int = 100,
-) -> typing.Tuple[np.ndarray, typing.Union[None, np.ndarray]]:
+    nb: int = 50,
+) -> typing.Tuple[np.ndarray, typing.Optional[np.ndarray]]:
     """
-    Construct a Hankel matrix using various methods with optional uncertainty calculations.
+    Construct a Hankel matrix with optional uncertainty calculations.
+
+    This function constructs a Hankel matrix using various methods, such as covariance-based
+    or data-driven approaches, and optionally computes uncertainty matrices.
 
     Parameters
     ----------
@@ -39,188 +42,152 @@ def build_hank(
     Yref : np.ndarray
         The reference data matrix with shape (number of reference channels, number of data points).
     br : int
-        The number of block rows to use in the Hankel matrix.
-    method : str
-        The method to use for constructing the Hankel matrix.
-        Options are 'cov_mm', 'cov_R', 'dat'ì
+        The number of block rows in the Hankel matrix.
+    method : {'cov', 'cov_R', 'dat'}
+        The method for constructing the Hankel matrix:
+        - 'cov': Covariance-based.
+        - 'cov_R': Covariance-based with auto-correlations.
+        - 'dat': Data-driven.
     calc_unc : bool, optional
-        Whether to calculate uncertainties (default is False). Only applicable for the 'cov_mm' method.
+        Whether to calculate uncertainty matrices. Default is False.
     nb : int, optional
-        The number of bootstrap samples to use for uncertainty calculations (default is 100).
+        The number of data blocks to use for uncertainty calculations. Default is 50.
 
     Returns
     -------
     Hank : np.ndarray
         The constructed Hankel matrix.
-    T : np.ndarray, optional
-        The uncertainty matrix. Only returned if `calc_unc` is True and `method` is 'cov_mm'.
+    T : np.ndarray or None
+        The uncertainty matrix, returned only if `calc_unc` is True. None otherwise.
 
     Raises
     ------
     AttributeError
-        If `calc_unc` is True but `method` is not 'cov_mm'.
-    AttributeError
-        If `method` is not one of 'cov_mm', 'cov_R', 'dat'ì
+        If `method` is not one of {'cov', 'cov_R', 'dat'}, or if data is insufficient for
+        uncertainty calculation.
 
     Notes
     -----
-    - The 'YfYp' method constructs separate future and past data matrices without combining them into a Hankel matrix.
-    - When `calc_unc` is True, the uncertainty matrix `T` is calculated using the bootstrap method, which quantifies the uncertainty of the Hankel matrix elements.
+    Uncertainty calculations follow the approach detailed in [DOME13]_.
     """
+
     Ndat = Y.shape[1]  # number of data points
-    l = Y.shape[0]  # number of chaiiels  # noqa E741 (ambiguous variable name 'l')
-    r = Yref.shape[0]  # number of reference chaiiels
-    p = int(br)  # number of block row (to use same notation as Dohler)
-    q = int(p + 1)  # block column
-    N = Ndat - p - q  # lenght of the Hankel matrix
+    l = Y.shape[0]  # number of channels # noqa E741 (ambiguous variable name 'l')
[…]
