+++ src/pyoma2/algorithms/fdd.py
@@ -389,12 +389,8 @@ class EFDD(FDD[EFDDRunParams, EFDDResult, typing.Iterable[float]]):
             If the algorithm has not been run and no results are available.
         """
 
-        if (
-            self.result is None
-            or self.result.Fn is None
-            or (isinstance(self.result, EFDDResult) and self.result.Xi is None)
-        ):
-            raise ValueError(f"Run {self.name} algorithm first and mpe method")
+        if not self.result:
+            raise ValueError("Run algorithm first")
 
         fig, ax = plot.EFDD_FIT_plot(
             Fn=self.result.Fn,
