+++ CHANGELOG.md
@@ -7,15 +7,30 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- uncertainty calculations for SSI algorithm
+- animation problem in pyvista
+- small fix (moved ax.grid()) in plt_data
+- updated docs
+
 ### Changed
 
 - Renamed `anim_mode_g2` to `anim_mode_geo2` in `GeometryMixin` class
 - Updated hierarchy for results and run_params classes
-- enhanced `fdd.plot_EFDDfit` error log
+- Renamed `plot_cluster()` method to `plot_freqvsdamp()`
+- SSI functions and classes re-organization:
+  - `cov_mm`method renamed to `cov`
+  - removed `ac2mp` function
+  - Hard criteria on MPC and MPD splitted
+  - HC on damping and on complex conjugate included into `SSI_poles`function
+  - order for run_param renamed to `order_in`
+  - Renamed uncertanties component from `xxx_cov`to `xxx_std`
 
 ### Added
 
 - pre commit in github workflow
+- clustering plotting functions to plot.py
 
 
 ## [1.0.0] - 2024-09-12
