### 2025-01-24 b7e0336
fix: pLSCF_MS algo; add: installation instructions for conda/mamba
Former-commit-id: 41a86e79daeb1da9dcc6854cd437dedeadc662ea [formerly 31bb0bd2a9bd4e188649f0336d9b2a16998b9f93] [formerly e05c0d16be46ffbdc89a785b6117c1d2b212c5e8 [formerly d15fa0d5193dcd88f4984dde3797b931690f7424]]
Former-commit-id: c30bae8f56d45174fc3bd0f0c1393668132afab3 [formerly 82562b106d6dca6f99f6916093c8f07ed508ce3b]
Former-commit-id: 6ec4609a9869287d3959b885f55f132da7fad63c
