+++ docs/Installation.rst
@@ -1,12 +1,17 @@
 Installation
 ============
 
-Install the library:
+Install the library with pip:
 
 .. code:: console
 
    pip install pyOMA-2
 
+or with conda/mamba:
+
+.. code:: console
+
+   conda install pyOMA-2
 
 You'll probably need to install **tk** for the GUI on your system, here some instructions:
 
