### 2025-01-24 d15fa0d
fix: pLSCF_MS algo; add: installation instructions for conda/mamba
