+++ docs/Example3 - Multisetup PoSER.rst
@@ -55,18 +55,18 @@ remains the same as described in the example for the single setup.
 .. code:: python
 
    # Initialise the algorithms for setup 1
-   ssicov1 = SSIcov(name="SSIcov1", method="cov_mm", br=50, ordmax=80)
+   ssicov1 = SSIcov(name="SSIcov1", method="cov", br=50, ordmax=80)
    # Add algorithms to the class
    ss1.add_algorithms(ssicov1)
    ss1.run_all()
 
    # Initialise the algorithms for setup 2
-   ssicov2 = SSIcov(name="SSIcov2", method="cov_mm", br=50, ordmax=80)
+   ssicov2 = SSIcov(name="SSIcov2", method="cov", br=50, ordmax=80)
    ss2.add_algorithms(ssicov2)
    ss2.run_all()
 
    # Initialise the algorithms for setup 2
-   ssicov3 = SSIcov(name="SSIcov3", method="cov_mm", br=50, ordmax=80)
+   ssicov3 = SSIcov(name="SSIcov3", method="cov", br=50, ordmax=80)
    ss3.add_algorithms(ssicov3)
    ss3.run_all()
 
