+++ tests/unit/functions/test_ssi.py
@@ -14,9 +14,9 @@ from tests.factory import assert_array_equal_with_nan
             np.array([[1, 2, 3, 4, 5]]),
             np.array([[1, 2, 3, 4, 5]]),
             1,
-            "cov_mm",
+            "cov",
             True,
-            np.array([[6.0, 4.0], [7.5, 5.0]]),
+            np.array([[10.0]]),
             None,
             False,
         ),
@@ -24,21 +24,11 @@ from tests.factory import assert_array_equal_with_nan
             np.array([[1, 2, 3, 4, 5]]),
             np.array([[1, 2, 3, 4, 5]]),
             1,
-            "cov_mm",
+            "cov",
             False,
-            np.array([[6.0, 4.0], [7.5, 5.0]]),
+            np.array([[10.0]]),
             True,
             False,
-        ),
-        (
-            np.array([[1, 2, 3, 4, 5]]),
-            np.array([[1, 2, 3, 4, 5]]),
-            1,
-            "cov_R",
-            True,
-            None,
-            None,
-            True,
         ),
         (
             np.array([[1, 2, 3, 4, 5]]),
@@ -46,7 +36,7 @@ from tests.factory import assert_array_equal_with_nan
             1,
             "cov_R",
             False,
-            np.array([[10.0, 11.0], [8.66666667, 10.0]]),
+            np.array([[13.75]]),
             True,
             False,
         ),
@@ -66,7 +56,7 @@ from tests.factory import assert_array_equal_with_nan
             1,
             "dat",
             False,
-            np.array([[2.82842712], [3.53553391]]),
+            np.array([[-3.65148372]]),
             True,
             False,
         ),
@@ -99,8 +89,7 @@ def test_build_hank(
                 Y=Y, Yref=Yref, br=br, method=method, calc_unc=calc_unc, nb=100
             )
             assert (
-                e.value
-                == "Uncertainty calculations are only available for 'cov_mm' method"
+                e.value == "Uncertainty calculations are only available for 'cov' method"
             )
     else:
         hank, uncertainty = ssi.build_hank(
@@ -124,32 +113,7 @@ def test_build_hank_invalid_method() -> None:
             calc_unc=False,
             nb=100,
         )
-        assert (
-            e.value == "Uncertainty calculations are only available for 'cov_mm' method"
-        )
-
-
-def test_ac2mp() -> None:
-    """Test the ac2mp function."""
-
-    A = np.array([[-1, -2], [3, -4]])
-    C = np.array([[1, 0], [0, 1]])
-    dt = 0.1
-
-    fn, xi, phi, *_ = ssi.ac2mp(A, C, dt)
-    assert fn.shape == (2,)
-    assert xi.shape == (2,)
-    assert phi.shape == (2, 2)
-
-    assert np.allclose(fn, np.array([4.35528095, 4.35528095]))
-    assert np.allclose(xi, np.array([-0.4207166, -0.4207166]))
-    assert np.allclose(
-        phi, np.array([[0.5 + 0.64549722j, 1.0 + 0.0j], [0.5 - 0.64549722j, 1.0 + 0.0j]])
-    )
-
-    assert isinstance(fn, np.ndarray)
-    assert isinstance(xi, np.ndarray)
-    assert isinstance(phi, np.ndarray)
[…]
