+++ .github/workflows/main.yml
@@ -17,7 +17,7 @@ jobs:
     - name: Set up Python
       uses: actions/setup-python@v5
       with:
-        python-version: "3.x"
+        python-version: "3.11"
     - name: Install pre-commit
       run: pip install pre-commit
     - name: Run pre-commit
