+++ requirements/requirementspy39+mac.txt
@@ -29,7 +29,6 @@ colorama==0.4.6; python_version < "3.13" and python_version >= "3.9"
 colorcet==3.1.0; python_version < "3.13" and python_version >= "3.9"
 comm==0.2.2; python_version < "3.13" and python_version >= "3.9"
 contourpy==1.3.0; python_version < "3.13" and python_version >= "3.9"
-coverage==7.6.1; python_version < "3.13" and python_version >= "3.9"
 coverage[toml]==7.6.1; python_version < "3.13" and python_version >= "3.9"
 cycler==0.12.1; python_version < "3.13" and python_version >= "3.9"
 debugpy==1.8.5; python_version < "3.13" and python_version >= "3.9"
@@ -65,7 +64,6 @@ jedi==0.19.1; python_version < "3.13" and python_version >= "3.9"
 jinja2==3.1.4; python_version < "3.13" and python_version >= "3.9"
 json5==0.9.25; python_version < "3.13" and python_version >= "3.9"
 jsonpointer==3.0.0; python_version < "3.13" and python_version >= "3.9"
-jsonschema==4.23.0; python_version < "3.13" and python_version >= "3.9"
 jsonschema-specifications==2023.12.1; python_version < "3.13" and python_version >= "3.9"
 jsonschema[format-nongpl]==4.23.0; python_version < "3.13" and python_version >= "3.9"
 jupyter-client==8.6.2; python_version < "3.13" and python_version >= "3.9"
@@ -133,7 +131,6 @@ pytest-cov==5.0.0; python_version < "3.13" and python_version >= "3.9"
 python-dateutil==2.9.0.post0; python_version < "3.13" and python_version >= "3.9"
 python-json-logger==2.0.7; python_version < "3.13" and python_version >= "3.9"
 pytz==2024.2; python_version < "3.13" and python_version >= "3.9"
-pyvista==0.44.1; python_version < "3.13" and python_version >= "3.9"
 pyvista[all]==0.44.1; python_version < "3.13" and python_version >= "3.9"
 pyvista[colormaps,io,jupyter]==0.44.1; python_version < "3.13" and python_version >= "3.9"
 pyvistaqt==0.11.1; python_version < "3.13" and python_version >= "3.9"
