+++ .github/workflows/main.yml
@@ -48,7 +48,7 @@ jobs:
       run: sudo apt-get install -y python3-tk
 
     - name: Install PDM
-      run: python -m pip install pdm==2.17.2
+      run: python -m pip install pdm==2.22.2
 
     - name: Install dependencies (Linux)
       if: matrix.os == 'ubuntu-latest'
