+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyOMA_2"
-version = "1.1.0"
+version = "1.1.1"
 description = "Python module for conducting Operational Modal Analysis"
 authors = [
     {name = "Dag Pasca", email = "dpa@treteknisk.no"},
