+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 
+## [1.1.1] - 2025-01-24
+
+- fix cluster_plt parameter name to Xi
+
 ## [1.1.0] - 2025-01-24
 
 ### Fixed
