+++ src/pyoma2/algorithms/data/result.py
@@ -5,7 +5,7 @@ related to the pyOMA2 module.
 
 from __future__ import annotations
 
-from typing import List, Optional, Union
+from typing import List, Optional
 
 import numpy as np
 import numpy.typing as npt
@@ -19,17 +19,13 @@ class BaseResult(BaseModel):
     Attributes
     ----------
     Fn : numpy.NDArray
-    Fn : numpy.NDArray
         Array of natural frequencies obtained from modal analysis.
-    Phi : numpy.NDArray
     Phi : numpy.NDArray
         Array of mode shape vectors obtained from modal analysis.
     """
 
     model_config = ConfigDict(from_attributes=True, arbitrary_types_allowed=True)
     # dopo mpe o mpe_from_plot
-    Fn: Optional[npt.NDArray[np.float64]] = None  # array of natural frequencies
-    Phi: Optional[npt.NDArray[np.float64]] = None  # array of Mode shape vectors
     Fn: Optional[npt.NDArray[np.float64]] = None  # array of natural frequencies
     Phi: Optional[npt.NDArray[np.float64]] = None  # array of Mode shape vectors
 
@@ -41,23 +37,15 @@ class FDDResult(BaseResult):
     Attributes
     ----------
     freq : numpy.NDArray
-    freq : numpy.NDArray
         Array of frequencies.
-    Sy : numpy.NDArray
     Sy : numpy.NDArray
         PSD obtained from the FDD analysis.
     S_val : numpy.NDArray
-    S_val : numpy.NDArray
         Singular values of the PSD.
-    S_vec : numpy.NDArray
     S_vec : numpy.NDArray
         Singular vectors of the PSD.
     """
 
-    freq: Optional[npt.NDArray[np.float64]] = None
-    Sy: Optional[npt.NDArray[np.float64]] = None
-    S_val: Optional[npt.NDArray[np.float64]] = None
-    S_vec: Optional[npt.NDArray[np.float64]] = None
     freq: Optional[npt.NDArray[np.float64]] = None
     Sy: Optional[npt.NDArray[np.float64]] = None
     S_val: Optional[npt.NDArray[np.float64]] = None
@@ -72,18 +60,13 @@ class EFDDResult(FDDResult):
     Attributes
     ----------
     freq : numpy.NDArray
-    freq : numpy.NDArray
         Array of frequencies.
-    Sy : numpy.NDArray
     Sy : numpy.NDArray
         PSD obtained from the analysis.
     S_val : numpy.NDArray
-    S_val : numpy.NDArray
         Singular values of the PSD.
     S_vec : numpy.NDArray
-    S_vec : numpy.NDArray
         Singular vectors of the PSD.
-    Xi : numpy.NDArray
     Xi : numpy.NDArray
         Array of damping ratios obtained from modal analysis.
     forPlot : list
@@ -91,8 +74,6 @@ class EFDDResult(FDDResult):
     """
 
     # dopo mpe, MPE_forPlot
-    Xi: Optional[npt.NDArray[np.float64]] = None  # array of damping ratios
-    forPlot: Optional[List] = None
     Xi: Optional[npt.NDArray[np.float64]] = None  # array of damping ratios
     forPlot: Optional[List] = None
 
@@ -104,53 +85,37 @@ class SSIResult(BaseResult):
     Attributes
     ----------
     Obs : numpy.NDArray, optional
-    Obs : numpy.NDArray, optional
         Observability matrix obtained from the SSI analysis.
-    A : list of numpy.NDArray, optional
     A : list of numpy.NDArray, optional
         List of system matrices A from the SSI analysis.
[…]
