+++ README.md
@@ -14,6 +14,9 @@ With this update, we've transformed pyOMA from a basic collection of functions i
 The module now supports analysis of both single and multi-setup data measurements, which includes handling multiple acquisitions with a mix of reference and roving sensors. We've also introduced interactive plots, allowing users to select desired modes for extraction directly from the plots generated by the algorithms. Additionally, a new feature enables users to define the geometry of the structures being tested, facilitating the visualization of mode shapes after modal results are obtained. The underlying functions of these classes have been rigorously revised, resulting in significant enhancements and optimizations.
 Since version 1.0.1, the uncertainty bounds of the modal properties can also be estimated for the SSI family of algorithms.
 
+**Clustering:**
+*We have introduced some new clustering specialised classes that allow the users to perform Automatic OMA. This update enables users to implement and compare a large number of the most popular algorithms introduced over the last 15 years, all within the same analysis framework. Furthermore, users have the flexibility to mix specific strategies from different algorithms to tailor a specific clustering process to their needs.*
+
 ## Documentation
 
 You can check the documentation at the following link:
