### 2025-01-31 dagghe/pyOMA2#26 @Nitnelav
JOSS Review : Fix the invalid DOI
https://github.com/openjournals/joss-reviews/issues/7656

The DOI `1016/j.ymssp.2013.01.012` is reported as invalid
