### 2025-01-31 dagghe/pyOMA2#27 @Nitnelav
JOSS Review : PDF formatting
https://github.com/openjournals/joss-reviews/issues/7656

The svg graph is not rendered properly in the PDF, the text is very small.
Also, the bullet points in the "Module’s structure." text are nor rendered properly in the PDF.
