### 2025-01-31 dagghe/pyOMA2#28 @Nitnelav
JOSS Review : Community Guidelines
https://github.com/openjournals/joss-reviews/issues/7656

The CONTRIBUTING.md file could start with less technical guidelines. For exemple : Are you welcoming any contributions ? Do you want to users to use Github issues or contact you directly ? etc.
