### 2025-01-31 dagghe/pyOMA2#29 @Nitnelav
JOSS Review : State of the field
https://github.com/openjournals/joss-reviews/issues/7656

There is no description on how this software compares to others.
