### 2025-01-31 #7656 @Nitnelav
Hello everyone,

I believe I have finished going through the reviewing checklist.

First of all I would like to thank the authors for making the effort to distribute their work as an open-source package meant to be used by the community.

I am not familiar with this field of structural analysis, but I enjoyed discovering the topic.
It was facilitated as the authors made a significant effort in making the software easy to use and understand by providing well written examples notebooks alongside exhaustive documentation of the API.

Though I would have liked to find more details in the JOSS paper. Especially after reading the [SoftwareX paper](https://www.sciencedirect.com/science/article/pii/S2352711022001340) (about pyOMA and pyOMA_gui) where I found way more context and descriptions about the first implementation of the software, most of which is probably still relevant to this version 2.

Apart from a few minor issues, the main blocking point for me is the absence of a "State of the field" on how this software compares to others.

I opened issues on the package GitHub repository for the various points I believe need to be addressed:

- Fix the invalid DOI: https://github.com/dagghe/pyOMA2/issues/26
- The SVG graph and some bullet poionts are not rendered properly in the PDF: https://github.com/dagghe/pyOMA2/issues/27
- The CONTRIBUTING.md file should start with more general guidelines: https://github.com/dagghe/pyOMA2/issues/28
- A clear "State of the field" is missing: https://github.com/dagghe/pyOMA2/issues/29

I checked everything accordingly the checklist comment.
