### 2025-02-06 dagghe/pyOMA2#27 @dagghe
Hi @Nitnelav thanks for the feedback! I will fix these problems in the paper in the coming days and also try to expand a bit on it as you suggested in the joss review issue page!
