### 2025-02-06 32e29ba
Update paper.bib
closes https://github.com/dagghe/pyOMA2/issues/26
