### 2025-02-06 44fead5
Update paper.bib
closes https://github.com/dagghe/pyOMA2/issues/26

Former-commit-id: 0e778b94cbeb9a550d5880785e88b7f3b5cec45b [formerly e5cfe47146dc05bf4e3d4aec46aeb600110726fb] [formerly 5f7de3555347c5cffd858c1df942ebdf30ffac28 [formerly 32e29bad29eefde965af3db4b4ff8bc17f56f91e]]
Former-commit-id: 0227caa73f9d2b6adc2c940d44c63d5aba765c19 [formerly c712ffeedd02973834628c272c9d24afc2906964]
Former-commit-id: 9ec3715e42d3436f10e51c1644f3a1c078d50129
