### 2025-02-06 dagghe/pyOMA2#26 @Nitnelav
I think it's because you forgot the "10."
