### 2025-02-06 dagghe/pyOMA2#29 @Nitnelav
Yes, well being fully free and open-source is a good enough "State of the field" (I consider Matlab stuff not free).
I believe it is written as such here : https://www.sciencedirect.com/science/article/pii/S2352711022001340#sec4

It just needs to appear in the JOSS paper
