### 2025-02-06 2bf98a1
docs: #28 edit contributing guidelines
Former-commit-id: 366adb47b8d946462596325cd3583ad69774edbf [formerly 0d67b2cbf4d7ab78c3fd0cf8695d23c6d03809aa] [formerly 5e26c8f2b1b7b67026515b8c29ed16bcc222aa1f [formerly 62b6b70e62f707eb50db3c43c7c0a21b6c9d30d8]]
Former-commit-id: 4eaa5dcb8a782b336b43c6d9eeb2a7411c479097 [formerly 5973c303cf54e38b75d2349b18df621c0f4cec1b]
Former-commit-id: 93d1108ed08eef5368401427c1686cb7643bb191
