+++ requirements/requirementspy38mac.txt
@@ -2,12 +2,12 @@
 # Please do not edit it manually.
 
 aiohappyeyeballs==2.4.0; python_version == "3.8"
-aiohttp==3.10.5; python_version == "3.8"
+aiohttp==3.10.11; python_version == "3.8"
 aiosignal==1.3.1; python_version == "3.8"
 alabaster==0.7.13; python_version == "3.8"
 annotated-types==0.7.0; python_version == "3.8"
 anyio==4.4.0; python_version == "3.8"
-appnope==0.1.4; sys_platform == "darwin" and python_version == "3.8" or platform_system == "Darwin" and python_version == "3.8"
+appnope==0.1.4; platform_system == "Darwin" and python_version == "3.8" or sys_platform == "darwin" and python_version == "3.8"
 argon2-cffi==23.1.0; python_version == "3.8"
 argon2-cffi-bindings==21.2.0; python_version == "3.8"
 arrow==1.3.0; python_version == "3.8"
@@ -18,7 +18,7 @@ attrs==24.2.0; python_version == "3.8"
 babel==2.16.0; python_version == "3.8"
 backcall==0.2.0; python_version == "3.8"
 beautifulsoup4==4.12.3; python_version == "3.8"
-bleach==6.1.0; python_version == "3.8"
+bleach[css]==6.1.0; python_version == "3.8"
 cachetools==5.5.0; python_version == "3.8"
 certifi==2024.8.30; python_version == "3.8"
 cffi==1.17.1; python_version == "3.8"
@@ -30,7 +30,6 @@ colorama==0.4.6; python_version == "3.8"
 colorcet==3.1.0; python_version == "3.8"
 comm==0.2.2; python_version == "3.8"
 contourpy==1.1.1; python_version == "3.8"
-coverage==7.6.1; python_version == "3.8"
 coverage[toml]==7.6.1; python_version == "3.8"
 cycler==0.12.1; python_version == "3.8"
 debugpy==1.8.5; python_version == "3.8"
@@ -42,7 +41,7 @@ et-xmlfile==1.1.0; python_version == "3.8"
 exceptiongroup==1.2.2; python_version == "3.8"
 executing==2.1.0; python_version == "3.8"
 fastjsonschema==2.20.0; python_version == "3.8"
-filelock==3.16.0; python_version == "3.8"
+filelock==3.16.1; python_version == "3.8"
 fonttools==4.53.1; python_version == "3.8"
 fqdn==1.5.1; python_version == "3.8"
 frozenlist==1.4.1; python_version == "3.8"
@@ -63,20 +62,19 @@ ipython==8.12.3; python_version == "3.8"
 ipywidgets==8.1.5; python_version == "3.8"
 isoduration==20.11.0; python_version == "3.8"
 jedi==0.19.1; python_version == "3.8"
-jinja2==3.1.4; python_version == "3.8"
+jinja2==3.1.5; python_version == "3.8"
 json5==0.9.25; python_version == "3.8"
 jsonpointer==3.0.0; python_version == "3.8"
-jsonschema==4.23.0; python_version == "3.8"
 jsonschema-specifications==2023.12.1; python_version == "3.8"
 jsonschema[format-nongpl]==4.23.0; python_version == "3.8"
-jupyter-client==8.6.2; python_version == "3.8"
+jupyter-client==8.6.3; python_version == "3.8"
 jupyter-core==5.7.2; python_version == "3.8"
 jupyter-events==0.10.0; python_version == "3.8"
 jupyter-lsp==2.2.5; python_version == "3.8"
 jupyter-server==2.14.2; python_version == "3.8"
[…]
