+++ requirements/requirementspy39+mac.txt
@@ -2,7 +2,7 @@
 # Please do not edit it manually.
 
 aiohappyeyeballs==2.4.0; python_version < "3.13" and python_version >= "3.9"
-aiohttp==3.10.5; python_version < "3.13" and python_version >= "3.9"
+aiohttp==3.11.12; python_version < "3.13" and python_version >= "3.9"
 aiosignal==1.3.1; python_version < "3.13" and python_version >= "3.9"
 alabaster==0.7.16; python_version < "3.13" and python_version >= "3.9"
 annotated-types==0.7.0; python_version < "3.13" and python_version >= "3.9"
@@ -17,7 +17,7 @@ async-timeout==4.0.3; python_version < "3.11" and python_version >= "3.9"
 attrs==24.2.0; python_version < "3.13" and python_version >= "3.9"
 babel==2.16.0; python_version < "3.13" and python_version >= "3.9"
 beautifulsoup4==4.12.3; python_version < "3.13" and python_version >= "3.9"
-bleach==6.1.0; python_version < "3.13" and python_version >= "3.9"
+bleach[css]==6.2.0; python_version < "3.13" and python_version >= "3.9"
 cachetools==5.5.0; python_version < "3.13" and python_version >= "3.9"
 certifi==2024.8.30; python_version < "3.13" and python_version >= "3.9"
 cffi==1.17.1; python_version < "3.13" and python_version >= "3.9"
@@ -40,7 +40,7 @@ et-xmlfile==1.1.0; python_version < "3.13" and python_version >= "3.9"
 exceptiongroup==1.2.2; python_version < "3.11" and python_version >= "3.9"
 executing==2.1.0; python_version < "3.13" and python_version >= "3.9"
 fastjsonschema==2.20.0; python_version < "3.13" and python_version >= "3.9"
-filelock==3.16.0; python_version < "3.13" and python_version >= "3.9"
+filelock==3.17.0; python_version < "3.13" and python_version >= "3.9"
 fonttools==4.53.1; python_version < "3.13" and python_version >= "3.9"
 fqdn==1.5.1; python_version < "3.13" and python_version >= "3.9"
 frozenlist==1.4.1; python_version < "3.13" and python_version >= "3.9"
@@ -61,26 +61,26 @@ ipython==8.18.1; python_version < "3.13" and python_version >= "3.9"
 ipywidgets==8.1.5; python_version < "3.13" and python_version >= "3.9"
 isoduration==20.11.0; python_version < "3.13" and python_version >= "3.9"
 jedi==0.19.1; python_version < "3.13" and python_version >= "3.9"
-jinja2==3.1.4; python_version < "3.13" and python_version >= "3.9"
+jinja2==3.1.5; python_version < "3.13" and python_version >= "3.9"
 json5==0.9.25; python_version < "3.13" and python_version >= "3.9"
 jsonpointer==3.0.0; python_version < "3.13" and python_version >= "3.9"
 jsonschema-specifications==2023.12.1; python_version < "3.13" and python_version >= "3.9"
 jsonschema[format-nongpl]==4.23.0; python_version < "3.13" and python_version >= "3.9"
-jupyter-client==8.6.2; python_version < "3.13" and python_version >= "3.9"
+jupyter-client==8.6.3; python_version < "3.13" and python_version >= "3.9"
 jupyter-core==5.7.2; python_version < "3.13" and python_version >= "3.9"
-jupyter-events==0.10.0; python_version < "3.13" and python_version >= "3.9"
+jupyter-events==0.12.0; python_version < "3.13" and python_version >= "3.9"
 jupyter-lsp==2.2.5; python_version < "3.13" and python_version >= "3.9"
[…]
