### 2025-02-10 fa282a0
added: docu files for clustering, new example
