### 2025-02-10 be3349a
added: docu files for clustering, new example
Former-commit-id: 793bbe3c53fa54a3cbe4db5f80d184323f77d69d [formerly 2c681043a2dfd227ab6afead129c9b4318a77cba] [formerly 8eef49cd8293597408b67812027e3b32b9911c57 [formerly fa282a01bff18eee40c839b514086e7b529e2efa]]
Former-commit-id: 7f12a40191430aa4a23c57e10895069eb0606e71 [formerly 96369f167798f2ee3cc244b43a959d2e9dbd1481]
Former-commit-id: 7ce21cfceb50aba00143aeda77f8411b93f1e148
