+++ docs/docu/2_4 Algo_autossi.rst
@@ -1,0 +1,21 @@
+The ``autossi`` algorithm module
+================================
+
+This module implements the automated version of the Stochastic Subspace Identification (SSI) algorithm.
+With this specialised class users can define clustering algorithms
+
+Classes:
+   :class:`.AutoSSI`
+      Implements the AutoSSI algorithm for single setup.
+
+.. Note::
+   Users should be familiar with the concepts of modal analysis and system identification to effectively use this module.
+
+
+The ``AutoSSI`` class
+-------------------------
+
+.. autoclass:: pyoma2.algorithms.autossi.AutoSSI
+   :members:
+   :inherited-members:
+   :show-inheritance:
