+++ CHANGELOG.md
@@ -20,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - run_params have been split in run_params and mpe_params
-
+- all dependencies are now mandatory (pyvista and openpyxl)
 
 ## [1.1.1] - 2025-01-24
 
