+++ docs/index.rst
@@ -30,7 +30,8 @@ Check out the project source_.
 
 
 .. important::
-   We have introduced some new clustering specialised classes that allow the users to perform Automatic OMA.
+
+   We have introduced some clustering specialised classes that allow the users to perform Automatic OMA.
    This update enables users to implement and compare a large number of the most popular algorithms
    introduced over the last 15 years, all within the same analysis framework. Furthermore, users have the
    flexibility to mix specific strategies from different algorithms to tailor a specific clustering process
