+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyOMA_2"
-version = "1.1.1"
+version = "1.2.0b1"
 description = "Python module for conducting Operational Modal Analysis"
 authors = [
     {name = "Dag Pasca", email = "dpa@treteknisk.no"},
@@ -16,6 +16,14 @@ dependencies = [
     "pydantic>=2.5.1",
     "tqdm>=4.66.1",
     "matplotlib>=3.7.4",
+    "pyvista[all]",
+    "pyvistaqt",
+    "PyQt5==5.15.10",
+    "PyQt5-Qt5==5.15.2; sys_platform != 'darwin'",
+    "PyQt5-Qt5==5.15.14; sys_platform == 'darwin'",
+    "PyQt5-sip==12.15.0",
+    "vtk==9.3.1; python_version != '3.8' or sys_platform != 'darwin'",
+    "openpyxl>=3.1.3",
 ]
 requires-python = ">=3.8,<3.13"
 readme = "README.md"
@@ -28,23 +36,6 @@ Documentation = "https://pyoma.readthedocs.io/en/main/"
 Repository = "https://github.com/dagghe/pyOMA2"
 Changelog = "https://github.com/dagghe/pyOMA2/blob/main/CHANGELOG.md"
 Contributing = "https://github.com/dagghe/pyOMA2/blob/main/CONTRIBUTING.md"
-
-
-[project.optional-dependencies]
-pyvista = [
-    "pyvista[all]",
-    "pyvistaqt",
-    "PyQt5==5.15.10",
-    "PyQt5-Qt5==5.15.2; sys_platform != 'darwin'",
-    "PyQt5-Qt5==5.15.14; sys_platform == 'darwin'",
-    "PyQt5-sip==12.15.0",
-    # # For macOS with Python 3.8, users will need to manually install the specific vtk wheel
-    # "vtk @ https://files.pythonhosted.org/packages/b3/15/40f8264f1b5379f12caf0e5153006a61c1f808937877c996e907610e8f23/vtk-9.3.1-cp38-cp38-macosx_10_10_x86_64.whl ; python_version == '3.8' and sys_platform == 'darwin'",
-    "vtk==9.3.1; python_version != '3.8' or sys_platform != 'darwin'",
-]
-openpyxl = [
-    "openpyxl>=3.1.3",
-]
 
 [build-system]
 requires = ["pdm-backend"]
