### 2025-02-10 3fac771
Merge with main: security patch
Former-commit-id: 34b7afbf306fb8263644879326982465c87f99ef [formerly ebff0b98bcc4933069b90ccd3aee0bd57a87f72d] [formerly c7be6a5fca33a8ad8770b74a69419235818b9062 [formerly ce2531a1df19a442124707666f4628f91920b5d2]]
Former-commit-id: de68467649f3dba17a675d2cb4fe0ba4dfb2ed77 [formerly 09b7eb3b3497908f19685040f5304ced7c4e8c1a]
Former-commit-id: 4cb28e5f8b686995486a72069a6a4f26960bbc4b
