+++ requirements/requirementspy39+mac.txt
@@ -1,9 +1,6 @@
 # This file is @generated by PDM.
 # Please do not edit it manually.
 
-aiohappyeyeballs==2.4.0; python_version < "3.13" and python_version >= "3.9"
-aiohttp==3.11.12; python_version < "3.13" and python_version >= "3.9"
-aiosignal==1.3.1; python_version < "3.13" and python_version >= "3.9"
 alabaster==0.7.16; python_version < "3.13" and python_version >= "3.9"
 annotated-types==0.7.0; python_version < "3.13" and python_version >= "3.9"
 anyio==4.4.0; python_version < "3.13" and python_version >= "3.9"
@@ -13,7 +10,6 @@ argon2-cffi-bindings==21.2.0; python_version < "3.13" and python_version >= "3.9
 arrow==1.3.0; python_version < "3.13" and python_version >= "3.9"
 asttokens==2.4.1; python_version < "3.13" and python_version >= "3.9"
 async-lru==2.0.4; python_version < "3.13" and python_version >= "3.9"
-async-timeout==4.0.3; python_version < "3.11" and python_version >= "3.9"
 attrs==24.2.0; python_version < "3.13" and python_version >= "3.9"
 babel==2.16.0; python_version < "3.13" and python_version >= "3.9"
 beautifulsoup4==4.12.3; python_version < "3.13" and python_version >= "3.9"
@@ -24,9 +20,7 @@ cffi==1.17.1; python_version < "3.13" and python_version >= "3.9"
 cfgv==3.4.0; python_version < "3.13" and python_version >= "3.9"
 chardet==5.2.0; python_version < "3.13" and python_version >= "3.9"
 charset-normalizer==3.3.2; python_version < "3.13" and python_version >= "3.9"
-cmocean==4.0.3; python_version < "3.13" and python_version >= "3.9"
 colorama==0.4.6; python_version < "3.13" and python_version >= "3.9"
-colorcet==3.1.0; python_version < "3.13" and python_version >= "3.9"
 comm==0.2.2; python_version < "3.13" and python_version >= "3.9"
 contourpy==1.3.0; python_version < "3.13" and python_version >= "3.9"
 coverage[toml]==7.6.1; python_version < "3.13" and python_version >= "3.9"
@@ -36,21 +30,18 @@ decorator==5.1.1; python_version < "3.13" and python_version >= "3.9"
 defusedxml==0.7.1; python_version < "3.13" and python_version >= "3.9"
 distlib==0.3.8; python_version < "3.13" and python_version >= "3.9"
 docutils==0.20.1; python_version < "3.13" and python_version >= "3.9"
-et-xmlfile==1.1.0; python_version < "3.13" and python_version >= "3.9"
 exceptiongroup==1.2.2; python_version < "3.11" and python_version >= "3.9"
 executing==2.1.0; python_version < "3.13" and python_version >= "3.9"
 fastjsonschema==2.20.0; python_version < "3.13" and python_version >= "3.9"
 filelock==3.17.0; python_version < "3.13" and python_version >= "3.9"
 fonttools==4.53.1; python_version < "3.13" and python_version >= "3.9"
 fqdn==1.5.1; python_version < "3.13" and python_version >= "3.9"
-frozenlist==1.4.1; python_version < "3.13" and python_version >= "3.9"
 ghp-import==2.1.0; python_version < "3.13" and python_version >= "3.9"
 h11==0.14.0; python_version < "3.13" and python_version >= "3.9"
 httpcore==1.0.5; python_version < "3.13" and python_version >= "3.9"
 httpx==0.27.2; python_version < "3.13" and python_version >= "3.9"
 identify==2.6.0; python_version < "3.13" and python_version >= "3.9"
[…]

[…]
