+++ docs/Example3 - Multisetup PoSER.rst
@@ -1,3 +1,4 @@
+=============================================================================
 Example3 - Multisetup with Post Separate Estimation Re-scaling (PoSER) method
 =============================================================================
 
@@ -8,151 +9,150 @@ using OpenSeesPy, and the corresponding Python script can be found `here <https:
 As always, first we import the necessary modules. All the files needed to run this
 example are available `here <https://github.com/dagghe/pyOMA-test-data/tree/main/test_data/3SL>`_.
 
-.. code:: python
+.. code-block:: python
 
-   import numpy as np
-   import pandas as pd
-   import matplotlib.pyplot as plt
+    import numpy as np
+    import pandas as pd
+    import matplotlib.pyplot as plt
 
-   from pyoma2.algorithms import SSIcov
-   from pyoma2.setup import MultiSetup_PoSER, SingleSetup
-   from pyoma2.support.utils.example_data import get_sample_data
+    from pyoma2.algorithms import SSIcov
+    from pyoma2.setup import MultiSetup_PoSER, SingleSetup
+    from pyoma2.support.utils.sample_data import get_sample_data
 
-For the **PoSER** approach, after importing the necessary modules and loading
-the data, the next step is to create a separate instance of the single setup
-class for each available dataset.
+For the **PoSER** approach, after importing the necessary modules and loading the data, the next step is to create a separate instance of the single setup class for each available dataset.
 
-In this example, we'll be working with a simulated dataset generated from a
-finite element model of a fictitious three-story, L-shaped building. This model
-was created using OpenSeesPy, and the corresponding Python script can be found
-in the GitHub repository.
+The exact natural frequencies of the system are: 2.63186, 2.69173, 3.43042, 8.29742, 8.42882, 10.6272, 14.0053, 14.093, 17.5741
 
-The exact natural frequencies of the system are:
-2.63186, 2.69173, 3.43042, 8.29742, 8.42882, 10.6272, 14.0053, 14.093, 17.5741
+.. code-block:: python
 
-.. code:: python
+    # import data files
+    set1 = np.load(get_sample_data(filename="set1.npy", folder="3SL"), allow_pickle=True)
+    set2 = np.load(get_sample_data(filename="set2.npy", folder="3SL"), allow_pickle=True)
+    set3 = np.load(get_sample_data(filename="set3.npy", folder="3SL"), allow_pickle=True)
 
-   # import data files
-   set1 = np.load(get_sample_data(filename="set1.npy", folder="3SL"), allow_pickle=True)
-   set2 = np.load(get_sample_data(filename="set2.npy", folder="3SL"), allow_pickle=True)
-   set3 = np.load(get_sample_data(filename="set3.npy", folder="3SL"), allow_pickle=True)
+    # create single setup
+    ss1 = SingleSetup(set1, fs=100)
+    ss2 = SingleSetup(set2, fs=100)
+    ss3 = SingleSetup(set3, fs=100)
 
-   # create single setup
-   ss1 = SingleSetup(set1, fs=100)
-   ss2 = SingleSetup(set2, fs=100)
-   ss3 = SingleSetup(set3, fs=100)
[…]
