+++ docs/Contributing.rst
@@ -1,0 +1,299 @@
+======================
+Contributing to pyOMA2
+======================
+
+Welcome! **pyOMA2** is a highly technical Python library focused on **Operational Modal Analysis (OMA)**. As such, contributions can come from both **programmers** (for bug fixes, best practices, and performance improvements) and **field experts** (for validation, methodology suggestions, and real-world applications). We appreciate any contributions that help improve the library!
+
+-------------------
+How to Get Involved
+-------------------
+
+We welcome contributions from both developers and OMA experts in different ways:
+
+- **Report Issues**: If you encounter a bug, have a feature request, or find something unclear in the documentation, please open a [GitHub Issue](https://github.com/dagghe/pyOMA2/issues).
+
+- **Code Contributions**: If you are a developer, you can help improve the codebase, fix bugs, enhance performance, or improve best practices by submitting a Pull Request (PR).
+
+- **Domain Expertise Contributions**: If you are an OMA specialist or engineer, we would love your input on methodology, validation of results, and possible improvements to the implemented techniques.
+
+- **Documentation Improvements**: If you notice unclear explanations or missing information, feel free to suggest improvements.
+
+- **Discussions & Questions**: If you're unsure about something, feel free to start a discussion in the Issues section.
+
+-----------------------
+Contribution Guidelines
+-----------------------
+
+Before making a contribution, please ensure you follow these best practices:
+
+- Follow the [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/) guidelines.
+- Ensure your changes do not break existing functionality.
+- Write tests for new features where applicable.
+- Be respectful and follow our community guidelines.
+
+-----
+Setup
+-----
+
+We use PDM as a dependency manager. Check the updated installation instructions from here, or follow these steps:
+
+----------------------
+Install PDM Linux/MAC
+----------------------
+
+.. code-block:: shell
+
+    curl -sSL https://pdm-project.org/install-pdm.py | python3 -
+
+-------------------
+Install PDM Windows
+-------------------
+
+.. code-block:: shell
+
+    (Invoke-WebRequest -Uri https://pdm-project.org/install-pdm.py -UseBasicParsing).Content | python -
+
+Add PATH to environment manager and then run the appropriate command to install all the dependencies based on your Python version and operating system:
+
+^^^^^^^^^^^^^^
+For Python 3.8
+^^^^^^^^^^^^^^
+
+**Linux**:
+
+.. code-block:: shell
+
+    pdm install --lockfile=pdm-py38unix.lock
+
+**Windows**:
+
+.. code-block:: shell
+
+    pdm install --lockfile=pdm-py38win.lock
+
+**macOS**:
+
+If you are using macOS with Python 3.8, you need to manually install the `vtk` package due to compatibility issues. You can do this by running the following command:
+
+.. code-block:: shell
+
+    pip install https://files.pythonhosted.org/packages/b3/15/40f8264f1b5379f12caf0e5153006a61c1f808937877c996e907610e8f23/vtk-9.3.1-cp38-cp38-macosx_10_10_x86_64.whl
+    pdm install --lockfile=pdm-py38macos.lock
+
[…]
