+++ CHANGELOG.md
@@ -12,10 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Extra1 example jupyter notebook
 - Example5 jupyter notebook for clustering
 - clustering.py function file
-- autossi.py for AutoSSI class
+- aoma.py for AutoSSI class
 - AutoSSIRunParams class, AutoSSIMPEParams class and AutoSSIResult class
 - Step1, Step2, Step3 and Clustering classes
-- documentation files for clustering
+- documentation files for clustering and examples
 
 ### Changed
 
