### 2025-02-11 63b02c7
fix: docstrings, joss paper.md issues #27 and #29
