+++ paper/paper.bib
@@ -220,3 +220,36 @@
   doi={10.1016/j.istruc.2023.04.115},
   url={https://doi.org/10.1016/j.istruc.2023.04.115},
 }
+
+@misc{solutions2001artemis,
+  title={ARTeMIS Extractor: Ambient Response Testing and Modal Identification Software, User’s Manual},
+  author={Solutions, Structural Vibration},
+  year={2001},
+  publisher={Demark}
+}
+
+@article{reynders2014macec,
+  title={MACEC 3.2: A Matlab toolbox for experimental and operational modal analysis},
+  author={Reynders, Edwin and Schevenels, Mattias and De Roeck, Guido},
+  journal={Department of Civil Engineering, KU Leuven},
+  year={2014}
+}
+
+@misc{koma,
+  author       = {Knut Andreas Kvåle},
+  title        = {KOMA: Knut's Operational Modal Analysis toolbox for Python},
+  year         = {2024},
+  url          = {https://github.com/knutankv/koma},
+  doi          = {10.5281/zenodo.14446286},
+  note         = {Version 1.3.6}
+}
+
+@article{pasca2022pyoma,
+  title={PyOMA and PyOMA\_GUI: a python module and software for operational modal analysis},
+  author={Pasca, Dag Pasquale and Aloisio, Angelo and Rosso, Marco Martino and Sotiropoulos, Stefanos},
+  journal={SoftwareX},
+  volume={20},
+  pages={101216},
+  year={2022},
+  publisher={Elsevier}
+}
