### 2025-02-11 fe3af08
fix: docstrings, joss paper.md issues #27 and #29
Former-commit-id: 20c0bac63528073a2f025d8e99d935b11d3dc54c [formerly 1c0f4008b9be5a528ad6c81a6f912de795afea42] [formerly 0c7cfde7b451acd1d91f14dc32ad70156ea6fee1 [formerly 63b02c7e093fc564f9a27447819ffce22f0937d2]]
Former-commit-id: fe81471a9371a9c65c653527448f8e11c9a75436 [formerly 3b3132c4b75e4f8800a5658e3e8294201fe7821b]
Former-commit-id: 1cdcf1bd2c09d369060c73930928f70a6834ff88
