+++ paper/paper.md
@@ -23,7 +23,7 @@ date: 12 September 2024
 bibliography: paper.bib
 ---
 
-# Summary
+## Summary
 
 Operational Modal Analysis (OMA) has garnered considerable attention from
 the engineering community in recent years and has established itself as
@@ -35,7 +35,8 @@ during the structure's normal operation. This makes OMA a more practical
 and efficient approach, as opposed to the traditional EMA, which requires
 both input and output data.
 
-# Statement of need
+## Statement of need
+
 `pyOMA2` is the latest and improved version of the pyOMA module, a Python
 library specifically designed for conducting operational modal analysis.
 It fully utilises Python's object-oriented capabilities to offer a
@@ -48,9 +49,11 @@ and visualising data. A key enhancement is the introduction of interactive
 plots, allowing users to select the desired modes directly from
 algorithm-generated graphs. Additionally, a new feature enables users to
 define the geometry of tested structures, facilitating the visualisation
-of mode shapes.
+of mode shapes. Since version 1.1.1 the uncertainty bounds of the modal
+properties can also be estimated for the SSI family of algorithms.
 
 The following algorithm are included in the module:
+
 - Frequency domain decomposition (FDD);
 - Enhanced frequency domain decomposition (EFDD);
 - Frequency spatial domain decomposition (FSDD);
@@ -67,6 +70,16 @@ Van Overschee and De Moor [@van2012subspace], Reynders [@reynders2012system],
 Rainieri and Fabbrocino [@rainieri2014operational], Amador [@amador2021robust],
 Döhler [@dohler2011subspace,dohler2013efficient,dohler2013uncertainty], and others.
 
+A few commercial software implements the algorithms mentioned above.
+The most known presumably are ARTeMIS [@solutions2001artemis], by Structural
+Vibration Solutions, and MACEC, a Matlab toolbox for modal testing and
+OMA [@reynders2014macec]. When it comes to open source modules the only ones
+available to the authors best knowledge are the first version of pyOMA
+[@pasca2022pyoma] and Koma [@koma], which is also an open-source Python library
+available on GitHub. It provides tools for OMA, focusing on simplicity and ease of use.
+Koma is designed to be a lightweight alternative to more general libraries like PyOMA,
+making it suitable for smaller projects or educational purposes.
+
 The module's reliability and applicability for research purposes have been
 demonstrated by the authors through various studies, such as
 [@alaggio2021two;@aloisio2020dynamic;@simoncelli2023intensity].
@@ -74,15 +87,17 @@ Additionally, the module has gained traction within the research community,
 as evidenced by its use in studies by
 [@saharan2023convolutional;@croce2023towards;@talebi2023interoperability;@abuodeh2023examining], and others.
 
-# Module's structure.
+## Module's structure
+
 The module is structured into three primary levels:
+
 1. At the first level are the `setup` classes. Users instantiate these classes by providing a data array and the sampling frequency for a single setup scenario, or a list of data arrays and their respective sampling frequencies, and reference indices, for a multi-setup scenario.
 2. The second level comprises the `algorithms` classes. Users can instantiate the algorithms they wish to run and then add them to the setup class.
 3. The third level contains the `support` classes, which serve as auxiliary components to the first two levels. This level includes various specialized classes:
-    * `result` classes, where outcomes are stored.
-    * `geometry` classes, for storing geometric data.
-    * `run_param` classes, where parameters used for running the algorithms are kept.
-    * Dedicated classes for animating mode shapes and interacting with plots generated by the algorithm classes.
+    - `result` classes, where outcomes are stored.
+    - `geometry` classes, for storing geometric data.
[…]
