### 2025-02-11 09a2189
add: info.pdf for joss paper
