### 2025-02-13 9202a1c
fix: tests and mpe run param class
Former-commit-id: f0bd4d58ed1dd92fc77f5a874ddae8a496671fee [formerly c40a7262d013f484a5fc6d59e2e3e47cc98ac69a] [formerly 3160ebbf92a1788088c38e1c45e6e0f9ccb91e1a [formerly 6781c482769e1dc709c031f1d24ad16f90a56879]]
Former-commit-id: 5a2115a1e85da90f2bd118e03c71d0d5fe064f9b [formerly f03921d925f8b869e77c6e523d15a1e5a91a928f]
Former-commit-id: 55bf923c98466cc24ebd16582bfc94e2d1020e77
