+++ requirements/requirementspy39+win.txt
@@ -1,75 +1,93 @@
 # This file is @generated by PDM.
 # Please do not edit it manually.
 
+aiohappyeyeballs==2.4.6; python_version < "3.13" and python_version >= "3.9"
+aiohttp==3.11.12; python_version < "3.13" and python_version >= "3.9"
+aiosignal==1.3.2; python_version < "3.13" and python_version >= "3.9"
 alabaster==0.7.16; python_version < "3.13" and python_version >= "3.9"
 annotated-types==0.7.0; python_version < "3.13" and python_version >= "3.9"
-anyio==4.4.0; python_version < "3.13" and python_version >= "3.9"
+anyio==4.8.0; python_version < "3.13" and python_version >= "3.9"
 argon2-cffi==23.1.0; python_version < "3.13" and python_version >= "3.9"
 argon2-cffi-bindings==21.2.0; python_version < "3.13" and python_version >= "3.9"
 arrow==1.3.0; python_version < "3.13" and python_version >= "3.9"
-asttokens==2.4.1; python_version < "3.13" and python_version >= "3.9"
+asttokens==3.0.0; python_version < "3.13" and python_version >= "3.9"
 async-lru==2.0.4; python_version < "3.13" and python_version >= "3.9"
-attrs==24.2.0; python_version < "3.13" and python_version >= "3.9"
-babel==2.16.0; python_version < "3.13" and python_version >= "3.9"
-beautifulsoup4==4.12.3; python_version < "3.13" and python_version >= "3.9"
+async-timeout==5.0.1; python_version < "3.11" and python_version >= "3.9"
+attrs==25.1.0; python_version < "3.13" and python_version >= "3.9"
+babel==2.17.0; python_version < "3.13" and python_version >= "3.9"
+beautifulsoup4==4.13.3; python_version < "3.13" and python_version >= "3.9"
 bleach[css]==6.2.0; python_version < "3.13" and python_version >= "3.9"
-cachetools==5.5.0; python_version < "3.13" and python_version >= "3.9"
-certifi==2024.8.30; python_version < "3.13" and python_version >= "3.9"
+cachetools==5.5.1; python_version < "3.13" and python_version >= "3.9"
+certifi==2025.1.31; python_version < "3.13" and python_version >= "3.9"
 cffi==1.17.1; python_version < "3.13" and python_version >= "3.9"
 cfgv==3.4.0; python_version < "3.13" and python_version >= "3.9"
 chardet==5.2.0; python_version < "3.13" and python_version >= "3.9"
-charset-normalizer==3.3.2; python_version < "3.13" and python_version >= "3.9"
+charset-normalizer==3.4.1; python_version < "3.13" and python_version >= "3.9"
+cmocean==4.0.3; python_version < "3.13" and python_version >= "3.9"
 colorama==0.4.6; python_version < "3.13" and python_version >= "3.9"
+colorcet==3.1.0; python_version < "3.13" and python_version >= "3.9"
 comm==0.2.2; python_version < "3.13" and python_version >= "3.9"
 contourpy==1.3.0; python_version < "3.13" and python_version >= "3.9"
-coverage[toml]==7.6.1; python_version < "3.13" and python_version >= "3.9"
+coverage[toml]==7.6.12; python_version < "3.13" and python_version >= "3.9"
 cycler==0.12.1; python_version < "3.13" and python_version >= "3.9"
-debugpy==1.8.5; python_version < "3.13" and python_version >= "3.9"
+debugpy==1.8.12; python_version < "3.13" and python_version >= "3.9"
 decorator==5.1.1; python_version < "3.13" and python_version >= "3.9"
 defusedxml==0.7.1; python_version < "3.13" and python_version >= "3.9"
[…]
