### 2025-02-13 6781c48
fix: tests and mpe run param class
