### 2025-02-13 4fb1ee1
fix: ssi_multi_setup error
