### 2025-02-16 dagghe/pyOMA2#31 @e-dub
JOSS review
Very nice work here.  After my initial interaction with the software and paper, I have made the following remarks

- [x] heavy git pull relative code size, nearly 200 MB for 5 MB of code. Is it pulling some old data in the .git folder? Maybe this can be reduced?
- [x] pip install slightly different name than code.  Can this be unified? pyoma2 vs pyoma-2
- [ ] Possible to provide non-ipython examples?

## Use
- [x] Can expected (theoretical) results be added to show correctness of results?
### Example1.ipynb:
- [x] Can be explained what we are looking at? Is this some kind of synthetic data or real measurements or the system parameters?
- [x] Maybe plot the data at beginning to bring it to life and understand what we are doing?
- [x] Please explain a bit what is going on, what is expected and why
- [x] Cell 1 is the same problem as rest of example?
- [x] maybe add "(downsampling)" to decimate. q=30 meaning every 30th sample?
- [x] analise -> analyze
- [x] instanciate -> instantiate, or initialize
- [x] Cell 4 gives several outputs with red background.  Is this desired behavior?
### Example2.ipynb:
- [x] include quick description of problem beyond provided reference may help new users (and this reviewer ;-)
- [x] several red messages when running.
- [x] error stating pyvista not installed despite being installed.  Using Ubuntu 24.10, tested with 3.11 and 3.12. Cell 20 properly displays figure.
### Example3.ipynb:
- [x] Same pyvista error as above for cells 7 and 9
### Example4.ipynb:
- [x] Same pyvista error as above for cell 5
### Example5.ipynb
- [x] Error in cell 4. AttributeError: 'NoneType' object has no attribute 'squeeze'
### Extra1.ipynb
- [x] Error in cell 14. IndexError: tuple index out of range

## Paper:
- [x] reference to original module pyOMA, i.e. 10.1016/j.softx.2022.101216
- [x] Expand to touch on deltas to pyOMA? Add need for pyOMA2 w.r.t. pyOMA?
- [x] Consider motivating work with possible uses and how this software meets gaps?
- [x] algorithm list as bulleted list
- [x] "the following algorithmS"
- [x] "key references" are listed without explaining why they are referenced. I recommend expanding this part.
- [x] users in community for pyOMA or pyOMA2? I recommend expanding to mention how used.
- [x] Number list for module primary levels?
- [x] asterisks needed in lines 48ff?
- [x] After reading, I do not fully understand what this module does and what input and output is.
