### 2025-02-16 dagghe/pyOMA2#31 @dfm88
> ## General
> * [ ]  heavy git pull relative code size, nearly 200 MB for 5 MB of code. Is it pulling some old data in the .git folder? Maybe this can be reduced?[ ]  pip install slightly different name than code.  Can this be unified? pyoma2 vs pyoma-2[ ]  Possible to provide non-ipython examples?

Dear @e-dub,

thank you for your time in reviewing this and to pointing out the issue. You are absolutely correct , the git history contains some large files that were committed and later removed in the past, leading to the current situation where cloning the repository downloads way more data than necessary.

@dagghe we should discuss whether to:
1. Remove these old large files from the history (this would require rewriting the entire git history)
2. Or alternatively, add a note in our contributing guidelines recommending contributors to use a shallow clone if they don't need the full history, so they wouldn't download the old huge files, with
```bash
git clone --depth=20 https://github.com/dagghe/pyOMA2.git
```

Please let us know if you have any preference between these two approaches.

Best regards
