### 2025-02-17 dagghe/pyOMA2#31 @dagghe
Hi @e-dub, thanks for the time spent reviewing our module! We will address your comments in the coming days.
