+++ src/pyoma2/functions/fdd.py
@@ -235,7 +235,7 @@ def SD_svalsvec(SD):
 # -----------------------------------------------------------------------------
 
 
-def FDD_MPE(
+def FDD_mpe(
     Sval,
     Svec,
     freq,
