### 2025-03-02 dagghe/pyOMA2#27 @faroit
@dagghe can this be closed?
