### 2025-03-09 dagghe/pyOMA2#29 @faroit
@dagghe could you integrate this into the paper? Let us know if you need some help on this issue
