### 2025-03-09 dagghe/pyOMA2#27 @faroit
@dagghe ^^
