### 2025-03-10 #7656 @e-dub
> [@e-dub](https://github.com/e-dub) can you give us a quick update on how the review is going?

@faroit please see [pyOMA2/issues/31](https://github.com/dagghe/pyOMA2/issues/31)
