### 2025-03-18 dagghe/pyOMA2#31 @dfm88
> ## General
> [x]  pip install slightly different name than code.  Can this be unified? pyoma2 vs pyoma-2

Unfortunately it is not possibile the name `pyOMA2` which was used for the very first release was wrongly deleted and PyPi does not allow to reuse the same name even if the repo was deleted, so we adopted `pyOMA-2`
