### 2025-03-22 dagghe/pyOMA2#27 @dagghe
Dear @faroit, sorry for the late reply, I have been out of office the last 2 weeks.
Yes, I think this issue can be closed.
