### 2025-03-22 dagghe/pyOMA2#29 @dagghe
@faroit I´ve added a sentence in the paper.md addressing the state of the field, but it has been commited to the cluster branch by mistake, I will add this to the main branch as well soon.
