### 2025-04-05 dagghe/pyOMA2#31 @faroit
@dagghe is there some update on when the issues are addressed?
