### 2025-04-05 dagghe/pyOMA2#27 @faroit
@Nitnelav please close if you think this has been sufficiently addressed
