### 2025-04-05 dagghe/pyOMA2#29 @faroit
@dagghe @Nitnelav was this sufficiently addressed?
