### 2025-04-05 dagghe/pyOMA2#27 @Nitnelav
Hello everyone. It looks like the changes were merged into the main branch instead of the Joss one... So the generated paper is still the old one
