### 2025-04-05 dagghe/pyOMA2#27 @dfm88
ok sorry @Nitnelav, I just aligned Joss with main
