### 2025-04-05 dagghe/pyOMA2#29 @Nitnelav
Yes ! It is addressed properly but the changes weren't merged into the Joss branch...
