### 2025-04-05 dagghe/pyOMA2#27 @Nitnelav
Mmmm it looks like the changes were not in the main branch initially... So they still do not appear in the Joss branch
