### 2025-04-14 dagghe/pyOMA2#31 @faroit
@dfm88 @dagghe can you update us here or tell us if you need more
