### 2025-04-14 dagghe/pyOMA2#27 @faroit
@dagghe do you need some help on this?
