### 2025-04-14 dagghe/pyOMA2#31 @dagghe
@faroit sorry I have been very busy with work!! I´ll do my best to address the issues during the easter holidays
