### 2025-04-14 dagghe/pyOMA2#31 @faroit
@dagghe thanks for the update.
