### 2025-04-14 dagghe/pyOMA2#27 @dfm88
I think this can be closed, just uploaded a png version of that image (ab43513) and changed all references
