### 2025-04-23 dagghe/pyOMA2#27 @Nitnelav
@dfm88 i'll have to disagree, the pdf still has some rendering issues around bullet points for exemple. Moreover, these changes are not the same as the ones done here : https://github.com/dagghe/pyOMA2/commit/fe3af0839474b45e9d4c7ad7ffe7813494f10221.

Can you get in touch with @dagghe so we can continue with the review process ? We've invested some time editing/reviewing this project and it is very close to being published. It would be quite unsatisfactory to not go through with the JOSS publication.
