### 2025-04-24 dagghe/pyOMA2#31 @dagghe
> ## Use
> ### General
>
>     [ ]  Can expected (theoretical) results be added to show correctness of results?
>

The exact result of the numerical simulations included in the examples are already reported in the files.

> ### Example1.ipynb:
>
>     [ ]  Can be explained what we are looking at? Is this some kind of synthetic data or real measurements or the system parameters?

The description of the example at hand has been reformulated and expanded.

>     [ ]  Maybe plot the data at beginning to bring it to life and understand what we are doing?

A plot of the data is already shown for example 2, which is a real dataset and thus shows a bit more interesting time histories. Here they the time histories would be white noises thus not so interesting/informative.

>     [ ]  Please explain a bit what is going on, what is expected and why

The description of the example at hand has been reformulated and expanded.

>     [ ]  Cell 1 is the same problem as rest of example?

I am not sure if I understand the question. In cell 1 we import the necessary modules and the example data (with the known modal parameters, since it is a numerical example).

>     [ ]  maybe add "(downsampling)" to decimate. q=30 meaning every 30th sample?

Even though the two terms are often used interchangeably there is a subtle difference: in downsampling one only reduce the sampling rate of a signal, while with decimation one first apply a low-pass filter to avoid aliasing and then downsample the filtered signal.

>     [ ]  Cell 4 gives several outputs with red background.  Is this desired behavior?

This is the expected behaviour, the outputs inform the users of what is going on and the progress of the analysis.

> ### Example2.ipynb:
>
>     [ ]  include quick description of problem beyond provided reference may help new users (and this reviewer ;-)

The description of the problem at hand has been reformulated and expanded.

>     [ ]  several red messages when running.

As explained above, this is the expected behaviour, the outputs inform the users of what is going on and the progress of the analysis.

>     [ ]  error stating pyvista not installed despite being installed.  Using Ubuntu 24.10, tested with 3.11 and 3.12. Cell 20 properly displays figure.
>
> ### Example3.ipynb:
>
>     * [ ]  Same pyvista error as above for cells 7 and 9
>
>
> ### Example4.ipynb:
>
>     * [ ]  Same pyvista error as above for cell 5

Could you give us a feedback if you still get this error message?

> ## Paper:
>
>     [ ]  reference to original module pyOMA, i.e. 10.1016/j.softx.2022.101216

We have included a reference to the old module.

>     [ ]  Expand to touch on deltas to pyOMA? Add need for pyOMA2 w.r.t. pyOMA?

We have modified the text, hopefully making more clear the improvements wrt to pyOMA.

>     [ ]  Consider motivating work with possible uses and how this software meets gaps?

Hopefully the modification made to the paper better explain these aspects.

>     [ ]  algorithm list as bulleted list
>     [ ]  "the following algorithmS"

Corrected

>     [ ]  "key references" are listed without explaining why they are referenced. I recommend expanding this part.

The section has been revised.

>     [ ]  users in community for pyOMA or pyOMA2? I recommend expanding to mention how used.

I am not sure I understand the question/comment.

>     [ ]  asterisks needed in lines 48ff?

Fixed.

>    [ ]  After reading, I do not fully understand what this module does and what input and output is.

I am sorry you feel this way; the software is highly technical and aimed at an audience primarily consisting of engineers working on in situ dynamic analysis and structural health monitoring. The input are the acceleration time histories acquired from a structure and the output are the modal properties of the structure under investigation.
