### 2025-04-24 dagghe/pyOMA2#27 @dagghe
Hi @Nitnelav hopefully the issues have been resolved now and the pdf should render properly
