### 2025-04-24 dagghe/pyOMA2#27 @Nitnelav
Ok looks good ! ping me when you think I should check everything one last time. Thanks
