### 2025-04-24 dagghe/pyOMA2#29 @dagghe
Dear @Nitnelav, I think this one can also be closed after the merging.
