### 2025-04-24 dagghe/pyOMA2#29 @Nitnelav
I agree !
