### 2025-04-29 dagghe/pyOMA2#31 @faroit
@e-dub can you please check the changes and close this issue if valid?
