### 2025-05-19 dagghe/pyOMA2#31 @faroit
@e-dub can you please check the changes and tell us if this addresses your original comments?
