### 2025-05-26 #7656 @faroit
@e-dub can you give us an update on your review? At JOSS we aim for reviews to be completed within about 2-4 weeks so this submission is overdue. As far as I can see there are a couple of points left:

- your main review was done here https://github.com/dagghe/pyOMA2/issues/31 and @e-dub would need to verify how the changes done by @dagghe in the meantime
- continue the checklist form https://github.com/openjournals/joss-reviews/issues/7656#issuecomment-2661425256
