+++ src/pyoma2/functions/ssi.py
@@ -25,9 +25,10 @@ def build_hank(
     Y: np.ndarray,
     Yref: np.ndarray,
     br: int,
-    method: typing.Literal["cov", "cov_R", "dat"],
+    method: typing.Literal["cov", "cov_R", "dat", "IOcov"],
     calc_unc: bool = False,
     nb: int = 50,
+    U: np.ndarray = None,
 ) -> typing.Tuple[np.ndarray, typing.Optional[np.ndarray]]:
     """
     Construct a Hankel matrix with optional uncertainty calculations.
@@ -113,7 +114,7 @@ def build_hank(
     elif method == "cov_R":
         # Correlations
         Ri = np.array(
-            [1 / (N) * np.dot(Y[:, k:], Yref[:, : Ndat - k].T) for k in range(p + q)]
+            [1 / (N - k) * np.dot(Y[:, k:], Yref[:, : Ndat - k].T) for k in range(p + q)]
         )
         # Assembling the Toepliz matrix
         Hank = np.vstack(
@@ -133,7 +134,7 @@ def build_hank(
                 for k in range(p + q):
                     print(f"{k=}, {p=}, {q=}")
                     res = np.array(
-                        [1 / (Nb) * np.dot(Y[:, : j * Nb - k], Yref[:, k : j * Nb].T)]
+                        [1 / (Nb - k) * np.dot(Y[:, : j * Nb - k], Yref[:, k : j * Nb].T)]
                     )
                     Ri = np.vstack([Ri, res])
                 Hcov_j = np.vstack(
@@ -182,10 +183,75 @@ def build_hank(
                 else:
                     T[:, j] = (Hdat_vec_j - Hvec0) / np.sqrt(nb * (nb - 1))
 
+    elif method == "IOcov":
+        try:
+            inp = U.shape[0]  # number of input channels
+        except AttributeError as e:
+            raise AttributeError(
+                "U must be provided when using method 'IOcov'."
+                "Please provide the input data matrix U."
+            ) from e
+        # preallocate
+        Uf = np.zeros(((p + 1) * inp, N))
+        Up = np.zeros((q * inp, N))
+        Yf = np.zeros(((p + 1) * l, N))
+        Yp = np.zeros((q * r, N))
+
+        # build past‐input blocks Up and past‐output blocks Yp
+        for i in range(q):
+            # take columns i .. i+N-1
+            Up[i * inp : (i + 1) * inp, :] = U[:, i : i + N]
+            Yp[i * r : (i + 1) * r, :] = Yref[:, i : i + N]
+
+        # build future‐input blocks Uf and future‐output blocks Yf
+        for i in range(p + 1):
+            Uf[i * inp : (i + 1) * inp, :] = U[:, q + i : q + i + N]
+            Yf[i * l : (i + 1) * l, :] = Y[:, q + i : q + i + N]
+
+        R2 = Yf.dot(Yp.T) / N
+        R4 = Yf.dot(Uf.T) / N
+        R8 = Uf.dot(Uf.T) / N
+        R5 = Yp.dot(Uf.T) / N
+
+        # Finally compute hk
+        Hank = R2 - R4 @ np.linalg.pinv(R8) @ R5.T
+
+        # Uncertainty calculations
+        logger.info("... calculating cov(H)...")
+        if calc_unc:
+            Nb = N // nb  # number of samples per segment
+            T = np.zeros(((p + 1) * q * l * r, nb))  # Square root of SIGMA_H
+            Hvec0 = Hank.reshape(-1, order="F")  # vectorized Hankel
+
+            for k in trange(nb):
+                Yf_k = Yf[:, (k * Nb) : ((k + 1) * Nb)]
[…]
