+++ src/pyoma2/algorithms/data/run_params.py
@@ -5,10 +5,10 @@ algorithms included in the pyOMA2 module.
 
 from __future__ import annotations
 
-import typing
 from typing import List, Literal, Optional, Tuple, Union
 
 import numpy as np
+import numpy.typing as npt
 from pydantic import BaseModel, ConfigDict, model_validator
 from typing_extensions import TypedDict
 
@@ -50,9 +50,9 @@ class SCDictType(TypedDict):
         The maximum thresholds for MAC difference.
     """
 
-    err_fn: float
-    err_xi: float
-    err_phi: float
+    err_fn: float = 0.05
+    err_xi: float = 0.1
+    err_phi: float = 0.05
 
 
 class BaseRunParams(BaseModel):
@@ -103,48 +103,6 @@ class EFDDRunParams(BaseRunParams):
     pov: float = 0.5
 
 
-class SSIRunParams(BaseRunParams):
-    """
-    Parameters for the Stochastic Subspace Identification (SSI) method.
-
-    Attributes
-    ----------
-    br : int
-        Number of block rows in the Hankel matrix.
-    method_hank : str or None, optional
-        Method used in the SSI algorithm. Options are ['data', 'cov', 'cov_R'].
-        Method used in the SSI algorithm. Options are ['data', 'cov', 'cov_R'].
-        Default is None.
-    ref_ind : list of int or None, optional
-        List of reference indices used for subspace identification. Default is None.
-    ordmin : int, optional
-        Minimum model order for the analysis. Default is 0.
-    ordmax : int or None, optional
-        Maximum model order for the analysis. Default is None.
-    step : int, optional
-        Step size for iterating through model orders. Default is 1.
-    sc : dict, optional
-        Soft validation criteria for the analysis.
-    hc : dict, optional
-        Hard validation criteria for the analysis.
-    calc_unc : bool, optional
-        Whether to calculate uncertainty. Default is False.
-    nb : int, optional
-        Number of bootstrap samples to use for uncertainty calculations (default is 50).
-    """
-
-    br: int = 20
-    method: str = None
-    ref_ind: Optional[List[int]] = None
-    ordmin: int = 0
-    ordmax: Optional[int] = None
-    step: int = 1
-    sc: SCDictType = dict(err_fn=0.05, err_xi=0.05, err_phi=0.05)
-    hc: HCDictType = dict(xi_max=0.1, mpc_lim=0.5, mpd_lim=0.5, CoV_max=0.05)
-    calc_unc: bool = False  # uncertainty calculations
-    nb: int = 50  # number of dataset blocks
-
-
 class pLSCFRunParams(BaseRunParams):
     """
     Parameters for the poly-reference Least Square Complex Frequency (pLSCF) method.
@@ -182,41 +140,48 @@ class pLSCFRunParams(BaseRunParams):
     hc: HCDictType = dict(xi_max=0.1, mpc_lim=0.7, mpd_lim=0.3)
 
 
-class AutoSSIRunParams(BaseRunParams):
+class SSIRunParams(BaseRunParams):
     """
-    Run parameters for automated SSI.
+    Parameters for the Stochastic Subspace Identification (SSI) method.
 
     Attributes
     ----------
     br : int
-        Number of block rows.
-    method : {'cov', 'cov_R', 'dat'}
-        Method for assembling the Hankel/subspace matrix.
-    ref_ind : list of int, optional
-        Indices of reference/projection channels.
-    ordmin : int
-        Minimum model order.
-    ordmax : int
-        Maximum model order.
-    step : int
-        Step size for model order increment.
-    calc_unc : bool
-        Whether to calculate uncertainty in results.
-    nb : int
-        Number of data blocks for uncertainty calculations.
+        Number of block rows in the Hankel matrix.
+    method_hank : str or None, optional
[…]
