+++ src/pyoma2/algorithms/__init__.py
@@ -1,5 +1,7 @@
 from .data.run_params import FDDRunParams, SSIRunParams, pLSCFRunParams  # noqa
 from .fdd import EFDD, EFDD_MS, FDD, FDD_MS, FSDD  # noqa
 from .plscf import pLSCF, pLSCF_MS  # noqa
-from .ssi import SSIcov, SSIcov_MS, SSIdat, SSIdat_MS  # noqa
+
+# from .ssi import SSIcov, SSIcov_MS, SSIdat, SSIdat_MS  # noqa
+from .ssi import SSI, SSI_MS  # noqa
 from .base import BaseAlgorithm  # noqa

[…]
