### 2025-06-03 dagghe/pyOMA2#31 @e-dub
> [@e-dub](https://github.com/e-dub) can you please check the changes and tell us if this addresses your original comments?

Please see checklist above. Example5 and Extra1 are showing errors when I run them and added new points accordingly.
