### 2025-06-04 32070c2
add: scikit-learn deps
