+++ requirements/requirementspy38unix.txt
@@ -62,6 +62,7 @@ ipywidgets==8.1.5; python_version == "3.8"
 isoduration==20.11.0; python_version == "3.8"
 jedi==0.19.1; python_version == "3.8"
 jinja2==3.1.5; python_version == "3.8"
+joblib==1.4.2; python_version == "3.8"
 json5==0.9.25; python_version == "3.8"
 jsonpointer==3.0.0; python_version == "3.8"
 jsonschema-specifications==2023.12.1; python_version == "3.8"
@@ -146,6 +147,7 @@ rfc3339-validator==0.1.4; python_version == "3.8"
 rfc3986-validator==0.1.1; python_version == "3.8"
 rich==13.8.1; python_version == "3.8"
 rpds-py==0.20.0; python_version == "3.8"
+scikit-learn==1.3.2; python_version == "3.8"
 scipy==1.10.1; python_version == "3.8"
 scooby==0.10.0; python_version == "3.8"
 send2trash==1.8.3; python_version == "3.8"
@@ -166,6 +168,7 @@ sphinxcontrib-qthelp==1.0.3; python_version == "3.8"
 sphinxcontrib-serializinghtml==1.1.5; python_version == "3.8"
 stack-data==0.6.3; python_version == "3.8"
 terminado==0.18.1; python_version == "3.8"
+threadpoolctl==3.5.0; python_version == "3.8"
 tinycss2==1.2.1; python_version == "3.8"
 tomli==2.2.1; python_version == "3.8"
 tornado==6.4.2; python_version == "3.8"
