+++ requirements/requirementspy39+unix.txt
@@ -149,6 +149,7 @@ send2trash==1.8.3; python_version < "3.13" and python_version >= "3.9"
 setuptools==74.1.2; python_version < "3.13" and python_version >= "3.9"
 simpervisor==1.0.0; python_version < "3.13" and python_version >= "3.9"
 six==1.16.0; python_version < "3.13" and python_version >= "3.9"
+sklearn==0.0.post12; python_version < "3.13" and python_version >= "3.9"
 sniffio==1.3.1; python_version < "3.13" and python_version >= "3.9"
 snowballstemmer==2.2.0; python_version < "3.13" and python_version >= "3.9"
 soupsieve==2.6; python_version < "3.13" and python_version >= "3.9"
