+++ requirements/requirementspy38mac.txt
@@ -153,6 +153,7 @@ send2trash==1.8.3; python_version == "3.8"
 setuptools==74.1.2; python_version == "3.8"
 simpervisor==1.0.0; python_version == "3.8"
 six==1.16.0; python_version == "3.8"
+sklearn==0.0.post12; python_version == "3.8"
 sniffio==1.3.1; python_version == "3.8"
 snowballstemmer==2.2.0; python_version == "3.8"
 soupsieve==2.6; python_version == "3.8"
