### 2025-06-04 4d263c4
add: python requirements
