+++ requirements/requirementspy39+mac.txt
@@ -63,7 +63,7 @@ isoduration==20.11.0; python_version < "3.13" and python_version >= "3.9"
 jedi==0.19.2; python_version < "3.13" and python_version >= "3.9"
 jinja2==3.1.5; python_version < "3.13" and python_version >= "3.9"
 joblib==1.5.1; python_version < "3.13" and python_version >= "3.9"
-json5==0.9.25; python_version < "3.13" and python_version >= "3.9"
+json5==0.10.0; python_version < "3.13" and python_version >= "3.9"
 jsonpointer==3.0.0; python_version < "3.13" and python_version >= "3.9"
 jsonschema-specifications==2024.10.1; python_version < "3.13" and python_version >= "3.9"
 jsonschema[format-nongpl]==4.23.0; python_version < "3.13" and python_version >= "3.9"
@@ -143,8 +143,8 @@ referencing==0.36.2; python_version < "3.13" and python_version >= "3.9"
 requests==2.32.3; python_version < "3.13" and python_version >= "3.9"
 rfc3339-validator==0.1.4; python_version < "3.13" and python_version >= "3.9"
 rfc3986-validator==0.1.1; python_version < "3.13" and python_version >= "3.9"
-rich==13.8.1; python_version < "3.13" and python_version >= "3.9"
-rpds-py==0.20.0; python_version < "3.13" and python_version >= "3.9"
+rich==13.9.4; python_version < "3.13" and python_version >= "3.9"
+rpds-py==0.22.3; python_version < "3.13" and python_version >= "3.9"
 scikit-learn==1.6.1; python_version < "3.13" and python_version >= "3.9"
 scipy==1.13.1; python_version < "3.13" and python_version >= "3.9"
 scooby==0.10.0; python_version < "3.13" and python_version >= "3.9"
@@ -167,7 +167,7 @@ sphinxcontrib-serializinghtml==2.0.0; python_version < "3.13" and python_version
 stack-data==0.6.3; python_version < "3.13" and python_version >= "3.9"
 terminado==0.18.1; python_version < "3.13" and python_version >= "3.9"
 threadpoolctl==3.6.0; python_version < "3.13" and python_version >= "3.9"
-tinycss2==1.3.0; python_version < "3.13" and python_version >= "3.9"
+tinycss2==1.4.0; python_version < "3.13" and python_version >= "3.9"
 tomli==2.2.1; python_version < "3.11" and python_version >= "3.9"
 tornado==6.4.2; python_version < "3.13" and python_version >= "3.9"
 tox==4.24.1; python_version < "3.13" and python_version >= "3.9"
