+++ requirements/requirementspy38unix.txt
@@ -63,7 +63,7 @@ isoduration==20.11.0; python_version == "3.8"
 jedi==0.19.2; python_version == "3.8"
 jinja2==3.1.5; python_version == "3.8"
 joblib==1.4.2; python_version == "3.8"
-json5==0.9.25; python_version == "3.8"
+json5==0.10.0; python_version == "3.8"
 jsonpointer==3.0.0; python_version == "3.8"
 jsonschema-specifications==2023.12.1; python_version == "3.8"
 jsonschema[format-nongpl]==4.23.0; python_version == "3.8"
@@ -133,7 +133,7 @@ pyqt5-sip==12.15.0; python_version == "3.8"
 pytest==8.3.4; python_version == "3.8"
 pytest-cov==5.0.0; python_version == "3.8"
 python-dateutil==2.9.0.post0; python_version == "3.8"
-python-json-logger==3.2.1; python_version == "3.8"
+python-json-logger==2.0.7; python_version == "3.8"
 pytz==2025.1; python_version == "3.8"
 pyvista[all]==0.44.2; python_version == "3.8"
 pyvista[colormaps,io,jupyter]==0.44.2; python_version == "3.8"
@@ -145,8 +145,8 @@ referencing==0.35.1; python_version == "3.8"
 requests==2.32.3; python_version == "3.8"
 rfc3339-validator==0.1.4; python_version == "3.8"
 rfc3986-validator==0.1.1; python_version == "3.8"
-rich==13.8.1; python_version == "3.8"
-rpds-py==0.20.0; python_version == "3.8"
+rich==13.9.4; python_version == "3.8"
+rpds-py==0.20.1; python_version == "3.8"
 scikit-learn==1.3.2; python_version == "3.8"
 scipy==1.10.1; python_version == "3.8"
 scooby==0.10.0; python_version == "3.8"
