+++ requirements/requirementspy38mac.txt
@@ -1,6 +1,9 @@
 # This file is @generated by PDM.
 # Please do not edit it manually.
 
+aiohappyeyeballs==2.4.0; python_version == "3.8"
+aiohttp==3.10.11; python_version == "3.8"
+aiosignal==1.3.1; python_version == "3.8"
 alabaster==0.7.13; python_version == "3.8"
 annotated-types==0.7.0; python_version == "3.8"
 anyio==4.4.0; python_version == "3.8"
@@ -10,6 +13,7 @@ argon2-cffi-bindings==21.2.0; python_version == "3.8"
 arrow==1.3.0; python_version == "3.8"
 asttokens==2.4.1; python_version == "3.8"
 async-lru==2.0.4; python_version == "3.8"
+async-timeout==4.0.3; python_version == "3.8"
 attrs==24.2.0; python_version == "3.8"
 babel==2.16.0; python_version == "3.8"
 backcall==0.2.0; python_version == "3.8"
@@ -21,7 +25,9 @@ cffi==1.17.1; python_version == "3.8"
 cfgv==3.4.0; python_version == "3.8"
 chardet==5.2.0; python_version == "3.8"
 charset-normalizer==3.3.2; python_version == "3.8"
+cmocean==4.0.3; python_version == "3.8"
 colorama==0.4.6; python_version == "3.8"
+colorcet==3.1.0; python_version == "3.8"
 comm==0.2.2; python_version == "3.8"
 contourpy==1.1.1; python_version == "3.8"
 coverage[toml]==7.6.1; python_version == "3.8"
@@ -31,18 +37,21 @@ decorator==5.1.1; python_version == "3.8"
 defusedxml==0.7.1; python_version == "3.8"
 distlib==0.3.8; python_version == "3.8"
 docutils==0.20.1; python_version == "3.8"
+et-xmlfile==1.1.0; python_version == "3.8"
 exceptiongroup==1.2.2; python_version == "3.8"
 executing==2.1.0; python_version == "3.8"
 fastjsonschema==2.20.0; python_version == "3.8"
 filelock==3.16.1; python_version == "3.8"
 fonttools==4.53.1; python_version == "3.8"
 fqdn==1.5.1; python_version == "3.8"
+frozenlist==1.4.1; python_version == "3.8"
 ghp-import==2.1.0; python_version == "3.8"
 h11==0.14.0; python_version == "3.8"
 httpcore==1.0.5; python_version == "3.8"
 httpx==0.27.2; python_version == "3.8"
 identify==2.6.0; python_version == "3.8"
 idna==3.8; python_version == "3.8"
+imageio==2.35.1; python_version == "3.8"
 imagesize==1.4.1; python_version == "3.8"
 importlib-metadata==8.5.0; python_version == "3.8"
 importlib-resources==6.4.5; python_version == "3.8"
@@ -50,6 +59,7 @@ iniconfig==2.0.0; python_version == "3.8"
 ipdb==0.13.13; python_version == "3.8"
 ipykernel==6.29.5; python_version == "3.8"
 ipython==8.12.3; python_version == "3.8"
+ipywidgets==8.1.5; python_version == "3.8"
 isoduration==20.11.0; python_version == "3.8"
 jedi==0.19.1; python_version == "3.8"
 jinja2==3.1.5; python_version == "3.8"
@@ -62,15 +72,23 @@ jupyter-core==5.7.2; python_version == "3.8"
 jupyter-events==0.10.0; python_version == "3.8"
 jupyter-lsp==2.2.5; python_version == "3.8"
 jupyter-server==2.14.2; python_version == "3.8"
+jupyter-server-proxy==4.4.0; python_version == "3.8"
[…]

[…]
