+++ src/pyoma2/functions/gen.py
@@ -905,152 +905,142 @@ def flatten_sns_names(sens_names, ref_ind=None) -> typing.List[str]:
 # -----------------------------------------------------------------------------
 
 
-def example_data() -> dict:
+def example_data() -> tuple:
     """
-    This function generates a time history of acceleration for a 5 DOF
-    system.
-
-    The function returns a (360001,5) array and a tuple containing: the
-    natural frequencies of the system (fn = (5,) array); the unity
-    displacement normalised mode shapes matrix (FI_1 = (5,5) array); and the
-    damping ratios (xi = float)
+    Simulate a 5‐DOF shear‐type building under white‐noise excitation,
+    with options for base‐ vs. DOF‐level inputs and choice of measurement type.
 
     Returns
     -------
-    acc : 2D array
-        Time histories of the 5 DOF of the system.
-    (fn, FI_1, xi) : tuple
-        Tuple containing the natural frequencies (fn), the mode shape
-        matrix (FI_1), and the damping ratio (xi) of the system.
-
+    Y_noisy : ndarray, shape (N, dof)
+        Simulated output time histories (for each DOF) of the requested measurement.
+    U : ndarray, shape (N, nu)
+        The actual input(s) (white noise + process noise) used.
+    (fn, xi, phi) : tuple
+        fn : array of natural frequencies [Hz], length = dof
+        xi : damping ratio (scalar)
+        phi: mass‐normalized mode‐shape matrix (dof × dof)
     """
+    # -------------------------
+    # Variable definition
+    dof = 5  # degrees of freedom
+    m_val = 25.91  # [kg]
+    k_val = 10000.0  # [N/m]
+    base_excitation: bool = False  # wheter to apply the input as base excitation or not
+    meas: str = "a"  # "d" = displacement, "v" = velocity, "a" = acceleration
+    N: int = 30001  # Number of samples
+    dt: float = 1 / 100  # Time increment [sec]
+    noise: float = 0.1  # Noise level (used for both process‐ and measurement noise)
+    xi: float = 0.02  # Damping ratio (all modes)
 
-    rng = np.random.RandomState(12345)  # Set the seed
-    fs = 600  # [Hz] Sampling freqiency
-    T = 900  # [sec] Period of the time series
+    # -------------------------
+    # Assembling mass matrix M and stiffness matrix K
+    M = np.eye(dof) * m_val
+    K = np.zeros((dof, dof))
+    for i in range(dof):
+        if i < dof - 1:
+            K[i, i] += k_val
+            K[i, i + 1] += -k_val
+            K[i + 1, i] += -k_val
+            K[i + 1, i + 1] += k_val
+        else:
+            K[i, i] += k_val
 
-    dt = 1 / fs  # [sec] time resolution
-    Ndat = int(T / dt)  # number of data points
+    # Eigen‐analysis: K φ = M φ Λ
+    lam, phi = linalg.eigh(K, M)
+    idx = np.argsort(lam)
+    lam = lam[idx]
+    phi = phi[:, idx]
 
-    t = np.linspace(0, T + dt, Ndat)
+    # Natural frequencies [Hz] and mass‐normalized mode shapes
+    w = np.sqrt(lam)  # [rad/s]
+    fn = w / (2 * np.pi)  # [Hz]
+    M_mod = phi.T @ M @ phi
+    norm_factors = 1.0 / np.sqrt(np.diag(M_mod))
[…]
