+++ src/pyoma2/functions/plot.py
@@ -24,11 +24,13 @@ from .gen import MAC
 logger = logging.getLogger(__name__)
 
 # =============================================================================
-# PLOT ALGORITMI
+# PLOT for CLUSTER
 # =============================================================================
 
 
-def plot_silhouette(distance_matrix, labels, name):
+def plot_silhouette(
+    distance_matrix: np.ndarray, labels: np.ndarray, name: str
+) -> typing.Tuple[plt.Figure, plt.Axes]:
     """
     Plot a silhouette plot for clustering results given a precomputed distance matrix.
 
@@ -38,50 +40,51 @@ def plot_silhouette(distance_matrix, labels, name):
         Pairwise distance matrix between samples.
     labels : array-like, shape (n_samples,)
         Cluster labels for each sample.
+    name : str
+        A label or title identifier used in the plot title.
 
     Returns
     -------
     fig : matplotlib.figure.Figure
         The figure object containing the silhouette plot.
+    ax : matplotlib.axes.Axes
+        The axes object containing the silhouette plot.
+
+    Notes
+    -----
+    - Computes the average silhouette score for the provided labels.
+    - Colors each cluster's silhouette values, and draws a vertical line at the average score.
     """
-    # Number of clusters
     labels = np.asarray(labels)
     n_clusters = len(np.unique(labels))
 
-    # Compute the average silhouette score
     silhouette_avg = silhouette_score(distance_matrix, labels, metric="precomputed")
-    # print(f"Average silhouette score: {silhouette_avg:.3f}")
-
-    # Compute silhouette values for each sample
     sample_silhouette_values = silhouette_samples(
         distance_matrix, labels, metric="precomputed"
     )
 
-    # Set up the plot
     fig, ax = plt.subplots(figsize=(10, 6))
     y_lower = 10
 
-    # Choose a colormap with enough distinct colors, excluding greys
-    if n_clusters <= 9:  # Exclude grey from tab10
+    # Choose colormap based on number of clusters, excluding greys where specified
+    if n_clusters <= 9:
         cmap = plt.get_cmap("tab10")
         colors = [cmap.colors[i] for i in range(len(cmap.colors)) if i != 7]
-    elif n_clusters <= 18:  # Exclude greys from tab20
+    elif n_clusters <= 18:
         cmap = plt.get_cmap("tab20")
         colors = [cmap.colors[i] for i in range(len(cmap.colors)) if i not in [14, 15]]
     else:
-        # Generate a colormap with n_labels distinct colors
         cmap = plt.cm.get_cmap("hsv", n_clusters)
         colors = cmap(np.linspace(0, 1, n_clusters))
 
-    # Plot silhouette for each cluster
-    for i in range(1, n_clusters):
-        # aggregate the silhouette scores for samples in cluster i, then sort
-        ith_vals = sample_silhouette_values[labels == i - 1]
+    for i in range(1, n_clusters + 1):
+        cluster_idx = i - 1
+        ith_vals = sample_silhouette_values[labels == cluster_idx]
         ith_vals.sort()
         size_i = ith_vals.shape[0]
         y_upper = y_lower + size_i
 
-        color = colors[i - 1]
+        color = colors[cluster_idx]
         ax.fill_betweenx(
             np.arange(y_lower, y_upper),
             0,
@@ -93,18 +96,16 @@ def plot_silhouette(distance_matrix, labels, name):
         ax.text(-0.05, y_lower + 0.5 * size_i, str(i))
         y_lower = y_upper + 10  # 10px gap between clusters
 
-    # Draw average silhouette score line with legend label
     avg_label = f"Avg = {silhouette_avg:.3f}"
     ax.axvline(
[…]
