### 2025-06-07 547abf4
fix: examples notebook
