+++ docs/index.rst
@@ -14,16 +14,16 @@ This is the new and updated version of **pyOMA** module, a Python module designe
 operational modal analysis. With this update, we've transformed **pyOMA** from a basic collection
 of functions into a more sophisticated module that fully leverages the capabilities of Python classes.
 
-The module now supports analysis of both single and multi-setup data measurements, which includes
-handling multiple acquisitions with a mix of reference and roving sensors. We've also introduced
-interactive plots, allowing users to select desired modes for extraction directly from the plots
-generated by the algorithms. Additionally, a new feature enables users to define the geometry of
-the structures being tested, facilitating the visualization of mode shapes after modal results are
-obtained. The underlying functions of these classes have been rigorously revised, resulting in
-significant enhancements and optimizations. Since version 1.0.1, the uncertainty bounds of the
-modal properties can also be estimated for the SSI family of algorithms.
+Key Features & Enhancements:
 
-We provide four :doc:`examples` to show the modules capabilities:
+- Support for single and multi-setup measurements, which includes handling multiple acquisitions with mixed reference and roving sensors.
+- Interactive plots for intuitive mode selection, users can now extract desired modes directly from algorithm-generated plots.
+- Structure geometry definition, enabling 3D visualization of mode shapes once modal results are obtained.
+- Uncertainty estimation for modal properties in Stochastic Subspace Identification (SSI) algorithms.
+- Specialized clustering classes for Automatic OMA using SSI, streamlining modal parameter extraction.
+- New OMAX (OMA with Exogenous Input) functionality for SSI, expanding the module’s capabilities to handle forced excitation scenarios.
+
+We provide five :doc:`examples` to show the modules capabilities:
 
 
 Check out the project source_.
