+++ README.md
@@ -1,4 +1,5 @@
 # pyOMA2
+
 ![pyoma2_logo_v2_COMPACT](https://github.com/dagghe/pyOMA2/assets/64746269/aa19bc05-d452-4749-a404-b702e6fe685d)
 
 [![python](https://img.shields.io/badge/python-3.8%20%7C%203.9%20%7C%203.10%20%7C%203.11%20%7C%203.12-blue.svg?style=flat&logo=python&logoColor=white)](https://www.python.org)
@@ -11,11 +12,14 @@ _______________________
 This is the new and updated version of pyOMA module, a Python module designed for conducting operational modal analysis.
 With this update, we've transformed pyOMA from a basic collection of functions into a more sophisticated module that fully leverages the capabilities of Python classes.
 
-The module now supports analysis of both single and multi-setup data measurements, which includes handling multiple acquisitions with a mix of reference and roving sensors. We've also introduced interactive plots, allowing users to select desired modes for extraction directly from the plots generated by the algorithms. Additionally, a new feature enables users to define the geometry of the structures being tested, facilitating the visualization of mode shapes after modal results are obtained. The underlying functions of these classes have been rigorously revised, resulting in significant enhancements and optimizations.
-Since version 1.0.1, the uncertainty bounds of the modal properties can also be estimated for the SSI family of algorithms.
+Key Features & Enhancements:
 
-**Clustering:**
-*We have introduced some new clustering specialised classes that allow the users to perform Automatic OMA. This update enables users to implement and compare a large number of the most popular algorithms introduced over the last 15 years, all within the same analysis framework. Furthermore, users have the flexibility to mix specific strategies from different algorithms to tailor a specific clustering process to their needs.*
+- Support for single and multi-setup measurements, which includes handling multiple acquisitions with mixed reference and roving sensors.
+- Interactive plots for intuitive mode selection, users can now extract desired modes directly from algorithm-generated plots.
+- Structure geometry definition, enabling 3D visualization of mode shapes once modal results are obtained.
+- Uncertainty estimation for modal properties in Stochastic Subspace Identification (SSI) algorithms.
+- Specialized clustering classes for Automatic OMA using SSI, streamlining modal parameter extraction.
+- New OMAX (OMA with Exogenous Input) functionality for SSI, expanding the module’s capabilities to handle forced excitation scenarios.
 
 ## Documentation
 
@@ -43,7 +47,6 @@ Windows:
 
 https://www.pythonguis.com/installation/install-tkinter-windows/
 
-
 Linux:
 
 https://www.pythonguis.com/installation/install-tkinter-linux/
@@ -55,6 +58,7 @@ https://www.pythonguis.com/installation/install-tkinter-mac/
 _____
 
 # Examples
+
 To see how the module works please take a look at the jupyter notebook provided:
 
 - [Example1 - Getting started.ipynb](Examples/Example1.ipynb)
@@ -63,9 +67,10 @@ To see how the module works please take a look at the jupyter notebook provided:
 - [Example4 - MultiSetup PreGER.ipynb](Examples/Example4.ipynb)
 - [Example5 - Clustering for Automatic OMA.ipynb](Examples/Example5.ipynb)
 - [Extra - Tips and Tricks 1.ipynb](Examples/Extra1.ipynb)
+
 _____
 
-# Schematic organisation of the module showing inheritance between classes.
+# Schematic organisation of the module showing inheritance between classes
 
 ![](docs/img/info.png "")
 
