### 2025-06-09 eef3e2a
refactor examples; refactor pyvista plotter
