+++ requirements/requirementspy38unix.txt
@@ -7,23 +7,23 @@ aiosignal==1.3.1; python_version == "3.8"
 alabaster==0.7.13; python_version == "3.8"
 annotated-types==0.7.0; python_version == "3.8"
 anyio==4.5.2; python_version == "3.8"
-argon2-cffi==23.1.0; python_version == "3.8"
+argon2-cffi==25.1.0; python_version == "3.8"
 argon2-cffi-bindings==21.2.0; python_version == "3.8"
 arrow==1.3.0; python_version == "3.8"
 asttokens==3.0.0; python_version == "3.8"
 async-lru==2.0.4; python_version == "3.8"
 async-timeout==5.0.1; python_version == "3.8"
-attrs==25.1.0; python_version == "3.8"
+attrs==25.3.0; python_version == "3.8"
 babel==2.17.0; python_version == "3.8"
 backcall==0.2.0; python_version == "3.8"
-beautifulsoup4==4.13.3; python_version == "3.8"
+beautifulsoup4==4.13.4; python_version == "3.8"
 bleach[css]==6.1.0; python_version == "3.8"
-cachetools==5.5.1; python_version == "3.8"
-certifi==2025.1.31; python_version == "3.8"
+cachetools==5.5.2; python_version == "3.8"
+certifi==2025.4.26; python_version == "3.8"
 cffi==1.17.1; python_version == "3.8"
 cfgv==3.4.0; python_version == "3.8"
 chardet==5.2.0; python_version == "3.8"
-charset-normalizer==3.4.1; python_version == "3.8"
+charset-normalizer==3.4.2; python_version == "3.8"
 cmocean==4.0.3; python_version == "3.8"
 colorama==0.4.6; python_version == "3.8"
 colorcet==3.1.0; python_version == "3.8"
@@ -31,22 +31,22 @@ comm==0.2.2; python_version == "3.8"
 contourpy==1.1.1; python_version == "3.8"
 coverage[toml]==7.6.1; python_version == "3.8"
 cycler==0.12.1; python_version == "3.8"
-debugpy==1.8.12; python_version == "3.8"
-decorator==5.1.1; python_version == "3.8"
+debugpy==1.8.14; python_version == "3.8"
+decorator==5.2.1; python_version == "3.8"
 defusedxml==0.7.1; python_version == "3.8"
 distlib==0.3.9; python_version == "3.8"
 docutils==0.20.1; python_version == "3.8"
 et-xmlfile==2.0.0; python_version == "3.8"
-exceptiongroup==1.2.2; python_version == "3.8"
+exceptiongroup==1.3.0; python_version == "3.8"
 executing==2.2.0; python_version == "3.8"
 fastjsonschema==2.21.1; python_version == "3.8"
 filelock==3.16.1; python_version == "3.8"
-fonttools==4.56.0; python_version == "3.8"
+fonttools==4.57.0; python_version == "3.8"
 fqdn==1.5.1; python_version == "3.8"
 frozenlist==1.5.0; python_version == "3.8"
 ghp-import==2.1.0; python_version == "3.8"
-h11==0.14.0; python_version == "3.8"
-httpcore==1.0.7; python_version == "3.8"
+h11==0.16.0; python_version == "3.8"
+httpcore==1.0.9; python_version == "3.8"
 httpx==0.28.1; python_version == "3.8"
 identify==2.6.1; python_version == "3.8"
 idna==3.10; python_version == "3.8"
@@ -54,30 +54,30 @@ imageio==2.35.1; python_version == "3.8"
 imagesize==1.4.1; python_version == "3.8"
 importlib-metadata==8.5.0; python_version == "3.8"
[…]
