+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "pyOMA_2"
-version = "1.2.0b2"
+version = "1.2.0b3"
 description = "Python module for conducting Operational Modal Analysis"
 authors = [
     {name = "Dag Pasca", email = "dpa@treteknisk.no"},
