### 2025-06-14 dagghe/pyOMA2#31 @faroit
@e-dub can you check if the issues are addressed sufficiently, please?
