### 2025-06-18 9f8a623
chore(docs): how to build docs
