### 2025-06-18 7b9b2d7
chore(docs): how to build docs
