### 2025-06-18 f1a3657
feat: stab plot color scheme selection for colorblind
