+++ src/pyoma2/functions/plot.py
@@ -24,6 +24,54 @@ from .gen import MAC
 logger = logging.getLogger(__name__)
 
 # =============================================================================
+# COLOR CONFIGURATION
+# =============================================================================
+
+DEFAULT_COLORS = {
+    "stable": "tab:green",
+    "unstable": "tab:red",
+}
+
+# Alternative color sets for different preferences
+ALTERNATIVE_COLORS = {
+    # Good for color blind
+    "classic": {"stable": "blue", "unstable": "orange"},
+    # High contrast for B&W printing
+    "high_contrast": {"stable": "black", "unstable": "gray"},
+    # Viridis colormap extremes
+    "viridis": {"stable": "#440154", "unstable": "#FDE725"},
+}
+
+
+def get_pole_colors(
+    color_scheme: typing.Literal[
+        "default", "classic", "high_contrast", "viridis"
+    ] = "default",
+) -> dict:
+    """
+    Get color scheme for stable and unstable poles.
+
+    Parameters
+    ----------
+    color_scheme : str, optional
+        Color scheme to use. Options: 'default', 'classic', 'high_contrast', 'viridis'.
+        Default is 'default' which uses colorblind-friendly colors.
+
+    Returns
+    -------
+    dict
+        Dictionary with 'stable' and 'unstable' color keys.
+    """
+    if color_scheme == "default":
+        return DEFAULT_COLORS.copy()
+    elif color_scheme in ALTERNATIVE_COLORS:
+        return ALTERNATIVE_COLORS[color_scheme].copy()
+    else:
+        logger.warning(f"Unknown color scheme '{color_scheme}', using default")
+        return DEFAULT_COLORS.copy()
+
+
+# =============================================================================
 # PLOT for CLUSTER
 # =============================================================================
 
@@ -978,6 +1026,9 @@ def stab_plot(
     Fn_std: typing.Optional[np.ndarray] = None,
     fig: typing.Optional[plt.Figure] = None,
     ax: typing.Optional[plt.Axes] = None,
+    color_scheme: typing.Literal[
+        "default", "classic", "high_contrast", "viridis"
+    ] = "default",
 ) -> typing.Tuple[plt.Figure, plt.Axes]:
     """
     Plot a stabilization chart of poles (frequency vs. model order) for stable vs unstable labeling.
@@ -1005,6 +1056,9 @@ def stab_plot(
         Existing figure to plot on. If None, a new figure is created.
     ax : matplotlib.axes.Axes, optional
         Existing axes to plot on. If None, new axes are created on the provided or new figure.
+    color_scheme : str, optional
+        Color scheme to use for stable/unstable poles. Options: 'default', 'classic',
+        'high_contrast', 'viridis'. Default is 'default' (colorblind-friendly).
 
     Returns
     -------
@@ -1015,11 +1069,14 @@ def stab_plot(
 
     Notes
     -----
-    - Green circles denote stable poles; red circles (if shown) denote unstable poles.
+    - By default uses colorblind-friendly blue/orange colors for stable/unstable poles.
     - Error bars represent frequency uncertainty if `Fn_std` is provided.
     """
     if fig is None and ax is None:
         fig, ax = plt.subplots(figsize=(8, 6), tight_layout=True)
+
+    # Get colors based on scheme
+    colors = get_pole_colors(color_scheme)
 
     # Extract stable and unstable frequencies
     Fns_stab = np.where(Lab == 1, Fn, np.nan)
@@ -1033,7 +1090,7 @@ def stab_plot(
     if hide_poles:
[…]
