### 2025-06-18 2ecc98e
feat: (#38) stab plot color scheme selection for colorblind
