+++ src/pyoma2/functions/plot.py
@@ -23,7 +23,55 @@ from .gen import MAC
 logger = logging.getLogger(__name__)
 
 # =============================================================================
-# PLOT ALGORITMI
+# COLOR CONFIGURATION
+# =============================================================================
+
+DEFAULT_COLORS = {
+    "stable": "tab:green",
+    "unstable": "tab:red",
+}
+
+# Alternative color sets for different preferences
+ALTERNATIVE_COLORS = {
+    # Good for color blind
+    "classic": {"stable": "blue", "unstable": "orange"},
+    # High contrast for B&W printing
+    "high_contrast": {"stable": "black", "unstable": "gray"},
+    # Viridis colormap extremes
+    "viridis": {"stable": "#440154", "unstable": "#FDE725"},
+}
+
+
+def get_pole_colors(
+    color_scheme: typing.Literal[
+        "default", "classic", "high_contrast", "viridis"
+    ] = "default",
+) -> dict:
+    """
+    Get color scheme for stable and unstable poles.
+
+    Parameters
+    ----------
+    color_scheme : str, optional
+        Color scheme to use. Options: 'default', 'classic', 'high_contrast', 'viridis'.
+        Default is 'default' which uses colorblind-friendly colors.
+
+    Returns
+    -------
+    dict
+        Dictionary with 'stable' and 'unstable' color keys.
+    """
+    if color_scheme == "default":
+        return DEFAULT_COLORS.copy()
+    elif color_scheme in ALTERNATIVE_COLORS:
+        return ALTERNATIVE_COLORS[color_scheme].copy()
+    else:
+        logger.warning(f"Unknown color scheme '{color_scheme}', using default")
+        return DEFAULT_COLORS.copy()
+
+
+# =============================================================================
+# PLOT for CLUSTER
 # =============================================================================
 
 
@@ -848,6 +896,9 @@ def stab_plot(
     Fn_std: np.array = None,
     fig: typing.Optional[plt.Figure] = None,
     ax: typing.Optional[plt.Axes] = None,
+    color_scheme: typing.Literal[
+        "default", "classic", "high_contrast", "viridis"
+    ] = "default",
 ) -> typing.Tuple[plt.Figure, plt.Axes]:
     """
     Plots a stabilization chart of the poles of a system.
@@ -874,6 +925,9 @@ def stab_plot(
         A matplotlib Axes object, by default None.
     Fn_std : np.ndarray, optional
         The covariance of the frequencies, used for error bars, by default None.
+    color_scheme : str, optional
+        Color scheme to use for stable/unstable poles. Options: 'default', 'classic',
+        'high_contrast', 'viridis'. Default is 'default' (colorblind-friendly).
 
     Returns
     -------
@@ -886,6 +940,10 @@ def stab_plot(
         fig, ax = plt.subplots(figsize=(8, 6), tight_layout=True)
 
     # Stable pole
+    # Get colors based on scheme
+    colors = get_pole_colors(color_scheme)
+
+    # Extract stable and unstable frequencies
     Fns_stab = np.where(Lab == 1, Fn, np.nan)
 
     # new or unstable
@@ -898,7 +956,7 @@ def stab_plot(
     if hide_poles:
         x = Fns_stab.flatten(order="F")
         y = np.array([i // len(Fns_stab) for i in range(len(x))]) * step
-        ax.plot(x, y, "go", markersize=7)
+        ax.plot(x, y, "o", color=colors["stable"], markersize=7)
 
         if Fn_std is not None:
             xerr = Fn_std.flatten(order="f")
@@ -912,8 +970,8 @@ def stab_plot(
[…]
