### 2025-07-31 #7656 @faroit
@dagghe please proceed with the items from the list [above](https://github.com/openjournals/joss-reviews/issues/7656#issuecomment-3141529109)
