### 2025-10-07 #7656 @faroit
@dagghe sorry for the slow response. I was sick for 2 weeks. It looks like we are almost ready. Did you see my review PR https://github.com/dagghe/pyOMA2/pull/39 ?
