### 2025-10-17 #7656 @dfm88
> [@dagghe](https://github.com/dagghe) sorry for the slow response. I was sick for 2 weeks. It looks like we are almost ready. Did you see my review PR [dagghe/pyOMA2#39](https://github.com/dagghe/pyOMA2/pull/39) ?

hi @faroit thanks for the pr, it was merged
