### 2025-11-05 #7656 @editorialbot
:wave: @openjournals/pe-eics, this paper is ready to be accepted and published.
