### 2025-11-11 #7656 @dfm88
> [@dagghe](https://github.com/dagghe) I've made some edits to the paper in [dagghe/pyOMA2#42](https://github.com/dagghe/pyOMA2/pull/42), please review and merge.
>
> Also, I noticed that one of the references seems to be missing information:
>
> > Reynders, E., Schevenels, M., & De Roeck, G. (2014). MACEC 3.2: A matlab toolbox for
> > experimental and operational modal analysis. Department of Civil Engineering, KU Leuven.
>
> Can you take a look at that one?

Thanks for pointing that out, I've updated the MACEC reference, it was listed as a journal article but it's actually a software toolbox manual, so I've changed it to the `@manual` entry type and added the organization details and URL
