### 2025-11-18 #7656 @faroit
@dagghe Happy to see this accepted and merged! This took much longer that I had hoped but I'm very pleased with the final result. @Nitnelav and @e-dub thanks for your reviewing!
